refactor(dev): split dep-graph.ts into focused modules
Extract the 3,262-line monolithic script into 18 focused modules under
scripts/dep-graph/. Key abstractions: withCache<T>() replaces 6 duplicate
cache patterns, callClaude<T>() replaces 4 duplicate CLI patterns, and
template files (HTML/CSS/JS) get real editor support instead of living
inside a template literal.
