+++ scripts/dep-graph.ts
@@ -7,7 +7,6 @@
  *   - Controller files       (controller -> service calls)
  *
  * Usage:
- * Usage:
  *   npx tsx scripts/dep-graph.ts              # domain-clustered layout with summaries (cached)
  *   npx tsx scripts/dep-graph.ts --json       # outputs raw JSON to stdout
  *   npx tsx scripts/dep-graph.ts --out dir    # writes HTML to dir/
@@ -16,3246 +15,6 @@
  *   npx tsx scripts/dep-graph.ts --publish    # publish to GitHub Pages (charliedowler/lightdash-dep-graph)
  */
 
-import * as fs from 'fs';
-import * as os from 'os';
-import * as path from 'path';
-import * as crypto from 'crypto';
-import * as ts from 'typescript';
-import { execSync } from 'child_process';
+import { main } from './dep-graph/main';
 
-// ---------------------------------------------------------------------------
-// Config
-// ---------------------------------------------------------------------------
-
-const BACKEND_SRC = path.resolve(__dirname, '../packages/backend/src');
-const SERVICE_REPO = path.join(
-    BACKEND_SRC,
-    'services/ServiceRepository.ts',
-);
-const CONTROLLERS_DIR = path.join(BACKEND_SRC, 'controllers');
-const ROUTERS_DIR = path.join(BACKEND_SRC, 'routers');
-const EE_DIR = path.join(BACKEND_SRC, 'ee');
-const EE_CONTROLLERS_DIR = path.join(EE_DIR, 'controllers');
-const EE_INDEX = path.join(EE_DIR, 'index.ts');
-
-// ---------------------------------------------------------------------------
-// Parsing
-// ---------------------------------------------------------------------------
-
-interface ServiceDeps {
-    models: string[];
-    clients: string[];
-    services: string[];
-}
-
-function parseServiceRepository(): Record<string, ServiceDeps> {
-    const src = fs.readFileSync(SERVICE_REPO, 'utf-8');
-
-    // Find each "public getXxx()" method boundary
-    const methodRe = /public\s+(get\w+)\(\)/g;
-    const starts: { pos: number; name: string }[] = [];
-    let m: RegExpExecArray | null;
-    while ((m = methodRe.exec(src)) !== null) {
-        starts.push({ pos: m.index, name: m[1] });
-    }
-
-    const services: Record<string, ServiceDeps> = {};
-
-    for (let i = 0; i < starts.length; i++) {
-        const block = src.slice(
-            starts[i].pos,
-            i + 1 < starts.length ? starts[i + 1].pos : src.length,
-        );
-
-        // Must have getService('key', ...) call
-        const keyMatch = block.match(/this\.getService\(\s*'(\w+)'/);
-        if (!keyMatch) continue;
-        const key = keyMatch[1];
-
-        const models = unique(
-            matchAll(block, /this\.models\.(get\w+)\(\)/g).map((x) =>
-                x.slice(3),
-            ),
-        );
-        const clients = unique(
-            matchAll(block, /this\.clients\.(get\w+)\(\)/g).map((x) =>
-                x.slice(3),
-            ),
-        );
-        const svcDeps = unique(
-            matchAll(block, /this\.(get\w+Service)\(\)/g)
-                .filter((x) => x !== 'getService')
-                .map((x) => x.slice(3)),
-        );
-
-        const displayName = key[0].toUpperCase() + key.slice(1);
-        services[displayName] = { models, clients, services: svcDeps };
-    }
-
-    return services;
-}
-
-function parseControllers(): Record<string, string[]> {
-    const controllers: Record<string, string[]> = {};
-
-    const patterns = [
-        path.join(CONTROLLERS_DIR, '*.ts'),
-        path.join(CONTROLLERS_DIR, 'v2', '*.ts'),
-    ];
-
-    for (const pattern of patterns) {
-        const dir = path.dirname(pattern);
-        if (!fs.existsSync(dir)) continue;
-
-        const files = fs.readdirSync(dir).filter((f) => f.endsWith('.ts'));
-        for (const file of files) {
-            if (
-                file === 'baseController.ts' ||
-                file === 'index.ts' ||
-                file.includes('.test.') ||
-                file.includes('.spec.')
-            )
-                continue;
-
-            const fpath = path.join(dir, file);
-            const content = fs.readFileSync(fpath, 'utf-8');
-
-            const calls = unique(
-                matchAll(content, /\.get(\w+Service)(?:<\w+>)?\(\)/g),
-            );
-
-            if (calls.length > 0) {
-                const isV2 = dir.endsWith('v2');
-                const name = file.replace('.ts', '');
-                const displayName = isV2 ? `v2/${name}` : name;
-                controllers[displayName] = calls.sort();
-            }
-        }
-    }
-
-    return controllers;
-}
-
-function parseRouters(): Record<string, string[]> {
-    const routers: Record<string, string[]> = {};
-
-    if (!fs.existsSync(ROUTERS_DIR)) return routers;
-
-    const files = fs.readdirSync(ROUTERS_DIR).filter((f) => f.endsWith('.ts'));
-    for (const file of files) {
-        if (
-            file === 'index.ts' ||
-            file.includes('.test.') ||
-            file.includes('.spec.')
-        )
-            continue;
-
-        const fpath = path.join(ROUTERS_DIR, file);
-        const content = fs.readFileSync(fpath, 'utf-8');
-
-        const calls = unique(
-            matchAll(content, /\.get(\w+Service)(?:<\w+>)?\(\)/g),
-        );
-
-        if (calls.length > 0) {
-            const name = file.replace('.ts', '');
-            routers[name] = calls.sort();
-        }
-    }
-
-    return routers;
-}
-
-interface EeParsed {
-    services: Record<string, ServiceDeps>;
-    modelNames: string[];
-    clientNames: string[];
-}
-
-function parseEeIndex(): EeParsed {
-    if (!fs.existsSync(EE_INDEX)) {
-        return { services: {}, modelNames: [], clientNames: [] };
-    }
-
-    const src = fs.readFileSync(EE_INDEX, 'utf-8');
-
-    const serviceBlockMatch = src.match(/serviceProviders:\s*\{([\s\S]*?)\n\s{8}\},/);
-    const modelBlockMatch = src.match(/modelProviders:\s*\{([\s\S]*?)\n\s{8}\},/);
-    const clientBlockMatch = src.match(/clientProviders:\s*\{([\s\S]*?)\n\s{8}\},/);
-
-    const services: Record<string, ServiceDeps> = {};
-
-    if (serviceBlockMatch) {
-        const block = serviceBlockMatch[1];
-        const providerRe = /^\s{12}(\w+):\s*\(/gm;
-        const providerStarts: { pos: number; name: string }[] = [];
-        let m: RegExpExecArray | null;
-        while ((m = providerRe.exec(block)) !== null) {
-            providerStarts.push({ pos: m.index, name: m[1] });
-        }
-
-        for (let i = 0; i < providerStarts.length; i++) {
-            const providerBlock = block.slice(
-                providerStarts[i].pos,
-                i + 1 < providerStarts.length ? providerStarts[i + 1].pos : block.length,
-            );
-            const key = providerStarts[i].name;
-
-            const models = unique(
-                matchAll(providerBlock, /models\.(get\w+)\(\)/g).map((x) => x.slice(3)),
-            );
-            const clients = unique(
-                matchAll(providerBlock, /clients\.(get\w+)\(\)/g).map((x) => x.slice(3)),
-            );
-            const svcDeps = unique(
-                matchAll(providerBlock, /repository\.(get\w+)\(\)/g).map((x) => x.slice(3)),
-            );
-
-            const displayName = key[0].toUpperCase() + key.slice(1);
-            services[displayName] = { models, clients, services: svcDeps };
-        }
-    }
-
-    const modelNames: string[] = [];
-    if (modelBlockMatch) {
-        const block = modelBlockMatch[1];
-        const nameRe = /^\s{12}(\w+):/gm;
-        let m: RegExpExecArray | null;
-        while ((m = nameRe.exec(block)) !== null) {
-            modelNames.push(m[1][0].toUpperCase() + m[1].slice(1));
-        }
-    }
-
-    const clientNames: string[] = [];
-    if (clientBlockMatch) {
-        const block = clientBlockMatch[1];
-        const nameRe = /^\s{12}(\w+):/gm;
-        let m: RegExpExecArray | null;
-        while ((m = nameRe.exec(block)) !== null) {
-            clientNames.push(m[1][0].toUpperCase() + m[1].slice(1));
-        }
-    }
-
-    return { services, modelNames, clientNames };
-}
-
-function parseEeControllers(): Record<string, string[]> {
-    const controllers: Record<string, string[]> = {};
-
-    if (!fs.existsSync(EE_CONTROLLERS_DIR)) return controllers;
-
-    const files = fs.readdirSync(EE_CONTROLLERS_DIR).filter((f) => f.endsWith('.ts'));
-    for (const file of files) {
-        if (
-            file === 'index.ts' ||
-            file.includes('.test.') ||
-            file.includes('.spec.')
-        )
-            continue;
-
-        const fpath = path.join(EE_CONTROLLERS_DIR, file);
-        const content = fs.readFileSync(fpath, 'utf-8');
-
-        const calls = unique(
-            matchAll(content, /\.get(\w+Service)(?:<\w+>)?\(\)/g),
-        );
-
-        if (calls.length > 0) {
-            const name = file.replace('.ts', '');
-            controllers[`ee/${name}`] = calls.sort();
-        }
-    }
-
-    return controllers;
-}
-
-function resolveFilePath(
-    id: string,
-    type: GraphNode['type'],
-    ee?: boolean,
-): string | undefined {
-    const candidates: string[] = [];
-    switch (type) {
-        case 'controller': {
-            if (ee || id.startsWith('ee/')) {
-                const stripped = id.replace(/^ee\//, '');
-                candidates.push(path.join(EE_CONTROLLERS_DIR, `${stripped}.ts`));
-            } else {
-                const stripped = id.replace(/^v2\//, '');
-                const sub = id.startsWith('v2/') ? 'v2/' : '';
-                candidates.push(
-                    path.join(CONTROLLERS_DIR, `${sub}${stripped}.ts`),
-                );
-            }
-            break;
-        }
-        case 'router':
-            candidates.push(path.join(ROUTERS_DIR, `${id}.ts`));
-            break;
-        case 'service':
-        case 'model':
-        case 'client': {
-            const dir =
-                type === 'service'
-                    ? 'services'
-                    : type === 'model'
-                      ? 'models'
-                      : 'clients';
-            if (ee) {
-                candidates.push(
-                    path.join(EE_DIR, dir, id, `${id}.ts`),
-                    path.join(EE_DIR, dir, `${id}.ts`),
-                );
-            }
-            candidates.push(
-                path.join(BACKEND_SRC, dir, id, `${id}.ts`),
-                path.join(BACKEND_SRC, dir, `${id}.ts`),
-            );
-            break;
-        }
-    }
-    return candidates.find((fp) => fs.existsSync(fp));
-}
-
-function resolveLineCount(
-    id: string,
-    type: GraphNode['type'],
-    ee?: boolean,
-): number | undefined {
-    const fp = resolveFilePath(id, type, ee);
-    if (fp) return fs.readFileSync(fp, 'utf-8').split('\n').length;
-    return undefined;
-}
-
-// ---------------------------------------------------------------------------
-// Complexity analysis (TypeScript AST)
-// ---------------------------------------------------------------------------
-
-interface Complexity {
-    cyclomatic: number;
-    cognitive: number;
-    maxFunctionCyclomatic: number;
-}
-
-function resolveComplexity(
-    id: string,
-    type: GraphNode['type'],
-    ee?: boolean,
-): Complexity | undefined {
-    const fp = resolveFilePath(id, type, ee);
-    if (!fp) return undefined;
-
-    const src = fs.readFileSync(fp, 'utf-8');
-    const sourceFile = ts.createSourceFile(fp, src, ts.ScriptTarget.Latest, true, ts.ScriptKind.TS);
-
-    let totalCyclomatic = 0;
-    let maxFunctionCyclomatic = 0;
-    let cognitive = 0;
-
-    function isFunctionLike(node: ts.Node): boolean {
-        return ts.isFunctionDeclaration(node) ||
-            ts.isFunctionExpression(node) ||
-            ts.isArrowFunction(node) ||
-            ts.isMethodDeclaration(node) ||
-            ts.isGetAccessor(node) ||
-            ts.isSetAccessor(node) ||
-            ts.isConstructorDeclaration(node);
-    }
-
-    function countCyclomaticInFunction(node: ts.Node): number {
-        let cc = 1;
-        function walk(n: ts.Node) {
-            switch (n.kind) {
-                case ts.SyntaxKind.IfStatement:
-                case ts.SyntaxKind.ForStatement:
-                case ts.SyntaxKind.ForInStatement:
-                case ts.SyntaxKind.ForOfStatement:
-                case ts.SyntaxKind.WhileStatement:
-                case ts.SyntaxKind.DoStatement:
-                case ts.SyntaxKind.CaseClause:
-                case ts.SyntaxKind.CatchClause:
-                case ts.SyntaxKind.ConditionalExpression:
-                    cc++;
-                    break;
-                case ts.SyntaxKind.BinaryExpression: {
-                    const bin = n as ts.BinaryExpression;
-                    if (bin.operatorToken.kind === ts.SyntaxKind.AmpersandAmpersandToken ||
-                        bin.operatorToken.kind === ts.SyntaxKind.BarBarToken ||
-                        bin.operatorToken.kind === ts.SyntaxKind.QuestionQuestionToken) {
-                        cc++;
-                    }
-                    break;
-                }
-            }
-            if (isFunctionLike(n) && n !== node) return;
-            ts.forEachChild(n, walk);
-        }
-        ts.forEachChild(node, walk);
-        return cc;
-    }
-
-    function walkForFunctions(node: ts.Node) {
-        if (isFunctionLike(node)) {
-            const cc = countCyclomaticInFunction(node);
-            totalCyclomatic += cc;
-            if (cc > maxFunctionCyclomatic) maxFunctionCyclomatic = cc;
-        }
-        ts.forEachChild(node, walkForFunctions);
-    }
-    walkForFunctions(sourceFile);
-
-    if (totalCyclomatic === 0) {
-        let topCC = 1;
-        function walkTop(n: ts.Node) {
-            switch (n.kind) {
-                case ts.SyntaxKind.IfStatement:
-                case ts.SyntaxKind.ForStatement:
-                case ts.SyntaxKind.ForInStatement:
-                case ts.SyntaxKind.ForOfStatement:
-                case ts.SyntaxKind.WhileStatement:
-                case ts.SyntaxKind.DoStatement:
-                case ts.SyntaxKind.CaseClause:
-                case ts.SyntaxKind.CatchClause:
-                case ts.SyntaxKind.ConditionalExpression:
-                    topCC++;
-                    break;
-                case ts.SyntaxKind.BinaryExpression: {
-                    const bin = n as ts.BinaryExpression;
-                    if (bin.operatorToken.kind === ts.SyntaxKind.AmpersandAmpersandToken ||
-                        bin.operatorToken.kind === ts.SyntaxKind.BarBarToken ||
-                        bin.operatorToken.kind === ts.SyntaxKind.QuestionQuestionToken) {
-                        topCC++;
-                    }
-                    break;
-                }
-            }
-            ts.forEachChild(n, walkTop);
-        }
-        ts.forEachChild(sourceFile, walkTop);
-        totalCyclomatic = topCC;
-        maxFunctionCyclomatic = topCC;
-    }
-
-    function isElseIf(node: ts.Node): boolean {
-        if (!ts.isIfStatement(node)) return false;
-        const parent = node.parent;
-        return !!parent && ts.isIfStatement(parent) && parent.elseStatement === node;
-    }
-
-    function isStructural(kind: ts.SyntaxKind): boolean {
-        return kind === ts.SyntaxKind.IfStatement ||
-            kind === ts.SyntaxKind.ForStatement ||
-            kind === ts.SyntaxKind.ForInStatement ||
-            kind === ts.SyntaxKind.ForOfStatement ||
-            kind === ts.SyntaxKind.WhileStatement ||
-            kind === ts.SyntaxKind.DoStatement ||
-            kind === ts.SyntaxKind.SwitchStatement ||
-            kind === ts.SyntaxKind.ConditionalExpression;
-    }
-
-    function cogWalk(node: ts.Node, depth: number) {
-        if (ts.isIfStatement(node)) {
-            if (isElseIf(node)) {
-                cognitive += 1;
-            } else {
-                cognitive += 1 + depth;
-            }
-            if (node.expression) cogWalk(node.expression, depth);
-            if (node.thenStatement) cogWalk(node.thenStatement, depth + 1);
-            if (node.elseStatement) {
-                if (ts.isIfStatement(node.elseStatement)) {
-                    cogWalk(node.elseStatement, depth);
-                } else {
-                    cognitive += 1;
-                    cogWalk(node.elseStatement, depth + 1);
-                }
-            }
-            return;
-        }
-
-        if (node.kind === ts.SyntaxKind.SwitchStatement) {
-            cognitive += 1 + depth;
-            ts.forEachChild(node, c => cogWalk(c, depth + 1));
-            return;
-        }
-
-        if (node.kind === ts.SyntaxKind.ForStatement ||
-            node.kind === ts.SyntaxKind.ForInStatement ||
-            node.kind === ts.SyntaxKind.ForOfStatement ||
-            node.kind === ts.SyntaxKind.WhileStatement ||
-            node.kind === ts.SyntaxKind.DoStatement) {
-            cognitive += 1 + depth;
-            ts.forEachChild(node, c => cogWalk(c, depth + 1));
-            return;
-        }
-
-        if (node.kind === ts.SyntaxKind.ConditionalExpression) {
-            cognitive += 1 + depth;
-            ts.forEachChild(node, c => cogWalk(c, depth));
-            return;
-        }
-
-        if (node.kind === ts.SyntaxKind.CatchClause) {
-            cognitive += 1;
-            ts.forEachChild(node, c => cogWalk(c, depth + 1));
-            return;
-        }
-
-        if (ts.isBinaryExpression(node)) {
-            const op = node.operatorToken.kind;
-            if (op === ts.SyntaxKind.AmpersandAmpersandToken ||
-                op === ts.SyntaxKind.BarBarToken ||
-                op === ts.SyntaxKind.QuestionQuestionToken) {
-                let parentOp: ts.SyntaxKind | undefined;
-                if (ts.isBinaryExpression(node.parent)) {
-                    parentOp = node.parent.operatorToken.kind;
-                }
-                if (parentOp !== op) {
-                    cognitive += 1;
-                }
-            }
-        }
-
-        if (isFunctionLike(node)) {
-            ts.forEachChild(node, c => cogWalk(c, depth + 1));
-            return;
-        }
-
-        ts.forEachChild(node, c => cogWalk(c, depth));
-    }
-
-    ts.forEachChild(sourceFile, c => cogWalk(c, 0));
-
-    return { cyclomatic: totalCyclomatic, cognitive, maxFunctionCyclomatic };
-}
-
-// ---------------------------------------------------------------------------
-// Graph building
-// ---------------------------------------------------------------------------
-
-interface RecentCommit {
-    hash: string;
-    message: string;
-    author: string;
-    relativeDate: string;
-}
-
-interface GitActivity {
-    commits: number;
-    authors: number;
-    churn: number;
-    recentCommits: RecentCommit[];
-}
-
-interface NodeDuplication {
-    clonedLines: number;
-    totalLines: number;
-    ratio: number;
-}
-
-interface SentryEndpoint {
-    route: string;
-    count: number;
-    p95Ms: number;
-    errorCount: number;
-}
-
-interface SentryActivity {
-    totalRequests: number;
-    totalErrors: number;
-    errorRate: number;
-    maxP95Ms: number;
-    endpoints: SentryEndpoint[];
-    spans: Array<{ name: string; count: number }>;
-    topError: string | null;
-    topErrorCount: number;
-    topErrorIssueId: string | null;
-    topErrorGroupId: string | null;
-    topErrorTransaction: string | null;
-}
-
-interface GraphNode {
-    id: string;
-    type: 'controller' | 'router' | 'service' | 'model' | 'client';
-    ee?: boolean;
-    domain?: string;
-    lineCount?: number;
-    complexity?: Complexity;
-    gitActivity?: GitActivity;
-    gitSummary?: string;
-    healthSummary?: string;
-    duplication?: NodeDuplication;
-    duplicationSummary?: string;
-    sentryActivity?: SentryActivity;
-}
-interface GraphEdge {
-    from: string;
-    to: string;
-    type: string;
-    duplication?: {
-        cloneCount: number;
-        totalLines: number;
-        advice?: string;
-        fragments: Array<{
-            lines: number;
-            fragment: string;
-            firstStart: number;
-            firstEnd: number;
-            secondStart: number;
-            secondEnd: number;
-        }>;
-    };
-}
-interface GraphData {
-    nodes: GraphNode[];
-    edges: GraphEdge[];
-    stats: {
-        controllers: number;
-        routers: number;
-        services: number;
-        models: number;
-        clients: number;
-        totalEdges: number;
-    };
-}
-
-function buildGraph(
-    services: Record<string, ServiceDeps>,
-    controllers: Record<string, string[]>,
-    routers: Record<string, string[]>,
-    eeData?: { services: Record<string, ServiceDeps>; controllers: Record<string, string[]>; modelNames: string[]; clientNames: string[] },
-): GraphData {
-    const nodes: GraphNode[] = [];
-    const edges: GraphEdge[] = [];
-    const nodeSet = new Set<string>();
-    const edgeSet = new Set<string>();
-
-    const addEdge = (from: string, to: string, type: string) => {
-        const key = `${from}\0${to}\0${type}`;
-        if (!edgeSet.has(key)) {
-            edgeSet.add(key);
-            edges.push({ from, to, type });
-        }
-    };
-
-    const addNode = (id: string, type: GraphNode['type'], ee?: boolean) => {
-        if (!nodeSet.has(id)) {
-            const lineCount = resolveLineCount(id, type, ee);
-            const complexity = resolveComplexity(id, type, ee);
-            nodes.push({ id, type, ...(ee && { ee }), ...(lineCount && { lineCount }), ...(complexity && { complexity }) });
-            nodeSet.add(id);
-        }
-    };
-
-    // Add controllers
-    for (const c of Object.keys(controllers).sort()) {
-        addNode(c, 'controller');
-    }
-
-    // Add routers
-    for (const r of Object.keys(routers).sort()) {
-        addNode(r, 'router');
-    }
-
-    // Add services
-    for (const s of Object.keys(services).sort()) {
-        addNode(s, 'service');
-    }
-
-    // Add models & clients from service deps
-    for (const deps of Object.values(services)) {
-        for (const m of deps.models) addNode(m, 'model');
-        for (const c of deps.clients) addNode(c, 'client');
-    }
-
-    // Controller -> Service edges
-    for (const [ctrl, svcs] of Object.entries(controllers)) {
-        for (const svc of svcs) {
-            if (nodeSet.has(svc)) {
-                addEdge(ctrl, svc, 'uses_service');
-            }
-        }
-    }
-
-    // Router -> Service edges
-    for (const [rtr, svcs] of Object.entries(routers)) {
-        for (const svc of svcs) {
-            if (nodeSet.has(svc)) {
-                addEdge(rtr, svc, 'router_uses_service');
-            }
-        }
-    }
-
-    // Service -> Model/Client/Service edges
-    for (const [svc, deps] of Object.entries(services)) {
-        for (const m of deps.models) {
-            addEdge(svc, m, 'injects_model');
-        }
-        for (const c of deps.clients) {
-            addEdge(svc, c, 'injects_client');
-        }
-        for (const s of deps.services) {
-            if (nodeSet.has(s)) {
-                addEdge(svc, s, 'injects_service');
-            }
-        }
-    }
-
-    // EE nodes and edges
-    if (eeData) {
-        for (const c of Object.keys(eeData.controllers).sort()) {
-            addNode(c, 'controller', true);
-        }
-
-        // EE services that don't already exist as OSS nodes are new EE-only services
-        for (const s of Object.keys(eeData.services).sort()) {
-            if (!nodeSet.has(s)) {
-                addNode(s, 'service', true);
-            }
-        }
-
-        // EE-only models and clients
-        for (const m of eeData.modelNames) {
-            if (!nodeSet.has(m)) addNode(m, 'model', true);
-        }
-        for (const c of eeData.clientNames) {
-            if (!nodeSet.has(c)) addNode(c, 'client', true);
-        }
-
-        // EE models/clients referenced as deps
-        for (const deps of Object.values(eeData.services)) {
-            for (const m of deps.models) {
-                if (!nodeSet.has(m)) addNode(m, 'model', true);
-            }
-            for (const c of deps.clients) {
-                if (!nodeSet.has(c)) addNode(c, 'client', true);
-            }
-        }
-
-        // EE controller -> service edges
-        for (const [ctrl, svcs] of Object.entries(eeData.controllers)) {
-            for (const svc of svcs) {
-                if (nodeSet.has(svc)) {
-                    addEdge(ctrl, svc, 'uses_service');
-                }
-            }
-        }
-
-        // EE service -> dep edges
-        for (const [svc, deps] of Object.entries(eeData.services)) {
-            for (const m of deps.models) {
-                if (nodeSet.has(m)) {
-                    addEdge(svc, m, 'injects_model');
-                }
-            }
-            for (const c of deps.clients) {
-                if (nodeSet.has(c)) {
-                    addEdge(svc, c, 'injects_client');
-                }
-            }
-            for (const s of deps.services) {
-                if (nodeSet.has(s)) {
-                    addEdge(svc, s, 'injects_service');
-                }
-            }
-        }
-    }
-
-    const modelCount = nodes.filter((n) => n.type === 'model').length;
-    const clientCount = nodes.filter((n) => n.type === 'client').length;
-
-    return {
-        nodes,
-        edges,
-        stats: {
-            controllers: Object.keys(controllers).length,
-            routers: Object.keys(routers).length,
-            services: Object.keys(services).length,
-            models: modelCount,
-            clients: clientCount,
-            totalEdges: edges.length,
-        },
-    };
-}
-
-// ---------------------------------------------------------------------------
-// Git activity
-// ---------------------------------------------------------------------------
-
-function collectGitActivity(nodes: GraphNode[], includeEe: boolean): void {
-    const repoRoot = path.resolve(__dirname, '..');
-    const relPathToNode = new Map<string, GraphNode>();
-    for (const n of nodes) {
-        const fp = resolveFilePath(n.id, n.type, n.ee);
-        if (fp) {
-            relPathToNode.set(path.relative(repoRoot, fp), n);
-        }
-    }
-
-    if (relPathToNode.size === 0) return;
-
-    const dirs = [
-        'packages/backend/src/services/',
-        'packages/backend/src/models/',
-        'packages/backend/src/clients/',
-        'packages/backend/src/controllers/',
-        'packages/backend/src/routers/',
-    ];
-
-    if (includeEe) {
-        dirs.push(
-            'packages/backend/src/ee/controllers/',
-            'packages/backend/src/ee/services/',
-            'packages/backend/src/ee/models/',
-            'packages/backend/src/ee/clients/',
-        );
-    }
-
-    let logOutput: string;
-    try {
-        logOutput = execSync(
-            `git log --format='COMMIT%x09%ae%x09%aI%x09%h%x09%s%x09%an%x09%ar' --name-only -- ${dirs.join(' ')}`,
-            { encoding: 'utf-8', maxBuffer: 50 * 1024 * 1024, cwd: repoRoot },
-        );
-    } catch {
-        console.warn('Warning: git log failed, skipping git activity.');
-        return;
-    }
-
-    const sixMonthsAgo = new Date();
-    sixMonthsAgo.setMonth(sixMonthsAgo.getMonth() - 6);
-
-    const commitCounts = new Map<string, number>();
-    const authorSets = new Map<string, Set<string>>();
-    const churnCounts = new Map<string, number>();
-    const recentCommitsMap = new Map<string, RecentCommit[]>();
-
-    let currentAuthor = '';
-    let currentDate = '';
-    let currentHash = '';
-    let currentMessage = '';
-    let currentAuthorName = '';
-    let currentRelDate = '';
-
-    for (const line of logOutput.split('\n')) {
-        if (line.startsWith('COMMIT\t')) {
-            const parts = line.split('\t');
-            currentAuthor = parts[1] || '';
-            currentDate = parts[2] || '';
-            currentHash = parts[3] || '';
-            currentMessage = parts[4] || '';
-            currentAuthorName = parts[5] || '';
-            currentRelDate = parts[6] || '';
-        } else if (line.trim() && !line.startsWith('COMMIT')) {
-            const file = line.trim();
-            if (!relPathToNode.has(file)) continue;
-
-            commitCounts.set(file, (commitCounts.get(file) || 0) + 1);
-
-            if (!authorSets.has(file)) authorSets.set(file, new Set());
-            authorSets.get(file)!.add(currentAuthor);
-
-            if (currentDate && new Date(currentDate) >= sixMonthsAgo) {
-                churnCounts.set(file, (churnCounts.get(file) || 0) + 1);
-            }
-
-            if (currentHash) {
-                const rc = recentCommitsMap.get(file) || [];
-                if (rc.length < 3) {
-                    rc.push({
-                        hash: currentHash,
-                        message: currentMessage,
-                        author: currentAuthorName,
-                        relativeDate: currentRelDate,
-                    });
-                    recentCommitsMap.set(file, rc);
-                }
-            }
-        }
-    }
-
-    for (const [relPath, node] of relPathToNode) {
-        const commits = commitCounts.get(relPath) || 0;
-        const authors = authorSets.get(relPath)?.size || 0;
-        const churn = churnCounts.get(relPath) || 0;
-        const recentCommits = recentCommitsMap.get(relPath) || [];
-        if (commits > 0) {
-            node.gitActivity = { commits, authors, churn, recentCommits };
-        }
-    }
-}
-
-// ---------------------------------------------------------------------------
-// Duplication analysis (jscpd)
-// ---------------------------------------------------------------------------
-
-const DUPLICATION_CACHE = path.join(__dirname, '.dep-graph-duplication.json');
-
-interface JscpdClone {
-    format: string;
-    lines: number;
-    fragment: string;
-    firstFile: { name: string; start: number; end: number; startLoc: { line: number }; endLoc: { line: number } };
-    secondFile: { name: string; start: number; end: number; startLoc: { line: number }; endLoc: { line: number } };
-}
-
-function computeDuplicationHash(graph: GraphData): string {
-    const stats: string[] = [];
-    for (const n of graph.nodes) {
-        const fp = resolveFilePath(n.id, n.type, n.ee);
-        if (fp && fs.existsSync(fp)) {
-            const st = fs.statSync(fp);
-            stats.push(`${fp}:${st.mtimeMs}:${st.size}`);
-        }
-    }
-    return crypto.createHash('sha256').update(stats.sort().join('\n')).digest('hex');
-}
-
-function runDuplicationAnalysis(graph: GraphData, force: boolean): void {
-    const hash = computeDuplicationHash(graph);
-    const shortHash = hash.slice(0, 12);
-
-    if (!force && fs.existsSync(DUPLICATION_CACHE)) {
-        try {
-            const cached = JSON.parse(fs.readFileSync(DUPLICATION_CACHE, 'utf-8'));
-            if (cached._hash === hash) {
-                console.log(`Duplication cache up to date (${shortHash}…). Use --refresh to rescan.`);
-                for (const edge of cached.edges) {
-                    graph.edges.push(edge);
-                }
-                for (const n of graph.nodes) {
-                    if (cached.perNode[n.id]) {
-                        n.duplication = cached.perNode[n.id];
-                    }
-                }
-                return;
-            }
-            console.log(`Source files changed since last duplication scan (${shortHash}…). Rescanning...`);
-        } catch {
-            console.log(`Duplication cache corrupt. Rescanning...`);
-        }
-    } else if (force) {
-        console.log(`Forced duplication rescan (${shortHash}…)...`);
-    } else {
-        console.log(`No duplication cache found (${shortHash}…). Scanning...`);
-    }
-
-    const fileToNode = new Map<string, GraphNode>();
-    for (const n of graph.nodes) {
-        const fp = resolveFilePath(n.id, n.type, n.ee);
-        if (fp) {
-            fileToNode.set(fp, n);
-        }
-    }
-
-    const tmpDir = fs.mkdtempSync(path.join(os.tmpdir(), 'dep-graph-jscpd-'));
-
-    const srcDirs = [
-        path.join(BACKEND_SRC, 'services'),
-        path.join(BACKEND_SRC, 'models'),
-        path.join(BACKEND_SRC, 'clients'),
-        path.join(BACKEND_SRC, 'controllers'),
-        path.join(BACKEND_SRC, 'routers'),
-        path.join(EE_DIR, 'services'),
-        path.join(EE_DIR, 'models'),
-        path.join(EE_DIR, 'clients'),
-        path.join(EE_DIR, 'controllers'),
-    ].filter(d => fs.existsSync(d));
-
-    console.log('Running jscpd for duplication analysis...');
-    try {
-        execSync(
-            `npx jscpd --reporters json --output ${escapeShellArg(tmpDir)} --format typescript --absolute --min-lines 5 --max-size 500kb --silent ${srcDirs.map(escapeShellArg).join(' ')}`,
-            { encoding: 'utf-8', maxBuffer: 50 * 1024 * 1024, cwd: path.resolve(__dirname, '..') },
-        );
-    } catch (err: any) {
-        if (!fs.existsSync(path.join(tmpDir, 'jscpd-report.json'))) {
-            console.warn('Warning: jscpd failed and produced no report. Skipping duplication analysis.');
-            fs.rmSync(tmpDir, { recursive: true });
-            return;
-        }
-    }
-
-    const reportPath = path.join(tmpDir, 'jscpd-report.json');
-    if (!fs.existsSync(reportPath)) {
-        console.warn('Warning: jscpd produced no report. Skipping duplication analysis.');
-        fs.rmSync(tmpDir, { recursive: true });
-        return;
-    }
-
-    const report = JSON.parse(fs.readFileSync(reportPath, 'utf-8'));
-    fs.rmSync(tmpDir, { recursive: true });
-
-    const duplicates: JscpdClone[] = report.duplicates || [];
-
-    const pairMap = new Map<string, {
-        cloneCount: number;
-        totalLines: number;
-        fragments: Array<{
-            lines: number;
-            fragment: string;
-            firstStart: number;
-            firstEnd: number;
-            secondStart: number;
-            secondEnd: number;
-        }>;
-    }>();
-
-    const perNodeDup = new Map<string, { clonedLines: number; totalLines: number }>();
-
-    for (const clone of duplicates) {
-        const nodeA = fileToNode.get(clone.firstFile.name);
-        const nodeB = fileToNode.get(clone.secondFile.name);
-
-        if (!nodeA || !nodeB) continue;
-        if (nodeA.id === nodeB.id) continue;
-
-        const pairKey = [nodeA.id, nodeB.id].sort().join('\0');
-        const existing = pairMap.get(pairKey) || { cloneCount: 0, totalLines: 0, fragments: [] };
-        existing.cloneCount++;
-        const lines = clone.lines || (clone.firstFile.endLoc.line - clone.firstFile.startLoc.line + 1);
-        existing.totalLines += lines;
-        if (existing.fragments.length < 5) {
-            const frag = (clone.fragment || '').slice(0, 300);
-            existing.fragments.push({
-                lines,
-                fragment: frag,
-                firstStart: clone.firstFile.startLoc.line,
-                firstEnd: clone.firstFile.endLoc.line,
-                secondStart: clone.secondFile.startLoc.line,
-                secondEnd: clone.secondFile.endLoc.line,
-            });
-        }
-        pairMap.set(pairKey, existing);
-    }
-
-    const newEdges: GraphEdge[] = [];
-    for (const [pairKey, data] of pairMap) {
-        const [fromId, toId] = pairKey.split('\0');
-        const edge: GraphEdge = {
-            from: fromId,
-            to: toId,
-            type: 'shares_code',
-            duplication: data,
-        };
-        newEdges.push(edge);
-        graph.edges.push(edge);
-    }
-
-    if (report.statistics?.formats?.typescript?.sources) {
-        for (const [filePath, sourceStats] of Object.entries(report.statistics.formats.typescript.sources) as Array<[string, any]>) {
-            const node = fileToNode.get(filePath);
-            if (node) {
-                const existing = perNodeDup.get(node.id) || { clonedLines: 0, totalLines: 0 };
-                existing.clonedLines += sourceStats.duplicatedLines || 0;
-                existing.totalLines += sourceStats.lines || 0;
-                perNodeDup.set(node.id, existing);
-            }
-        }
-    }
-
-    const perNodeResult: Record<string, NodeDuplication> = {};
-    for (const [nodeId, stats] of perNodeDup) {
-        const ratio = stats.totalLines > 0 ? stats.clonedLines / stats.totalLines : 0;
-        const dup: NodeDuplication = {
-            clonedLines: stats.clonedLines,
-            totalLines: stats.totalLines,
-            ratio,
-        };
-        perNodeResult[nodeId] = dup;
-        const node = graph.nodes.find(n => n.id === nodeId);
-        if (node) {
-            node.duplication = dup;
-        }
-    }
-
-    console.log(`Found ${duplicates.length} clone pairs, ${newEdges.length} cross-node duplication edges.`);
-
-    const cacheData = {
-        _hash: hash,
-        _generatedAt: new Date().toISOString(),
-        edges: newEdges,
-        perNode: perNodeResult,
-    };
-    fs.writeFileSync(DUPLICATION_CACHE, JSON.stringify(cacheData, null, 2) + '\n');
-    console.log(`Cached to ${DUPLICATION_CACHE}`);
-}
-
-// ---------------------------------------------------------------------------
-// Duplication summary (via Claude CLI)
-// ---------------------------------------------------------------------------
-
-const DUPLICATION_SUMMARY_CACHE = path.join(__dirname, '.dep-graph-duplication-summaries.json');
-
-function summarizeDuplication(graph: GraphData, force: boolean): void {
-    const dupEdges = graph.edges.filter(e => e.type === 'shares_code' && e.duplication);
-    const dupNodes = graph.nodes.filter(n => n.duplication && n.duplication.ratio > 0);
-
-    if (dupEdges.length === 0 && dupNodes.length === 0) {
-        console.log('No duplication data to summarize.');
-        return;
-    }
-
-    const parts = [
-        ...dupEdges.map(e => `${e.from}\0${e.to}\0${e.duplication!.cloneCount}\0${e.duplication!.fragments.map(f => f.fragment).join('')}`),
-        ...dupNodes.map(n => `${n.id}\0${n.duplication!.clonedLines}\0${n.duplication!.totalLines}`),
-    ].sort();
-    const hash = crypto.createHash('sha256').update(parts.join('\n')).digest('hex');
-    const shortHash = hash.slice(0, 12);
-
-    if (!force && fs.existsSync(DUPLICATION_SUMMARY_CACHE)) {
-        try {
-            const cached = JSON.parse(fs.readFileSync(DUPLICATION_SUMMARY_CACHE, 'utf-8'));
-            if (cached._hash === hash) {
-                console.log(`Duplication summary cache up to date (${shortHash}…). Use --refresh to regenerate.`);
-                for (const edge of dupEdges) {
-                    const sorted = [edge.from, edge.to].sort();
-                    const advice = cached.pairAdvice[sorted.join('\0')] || cached.pairAdvice[sorted.join(' ')];
-                    if (advice) {
-                        edge.duplication!.advice = advice;
-                    }
-                }
-                for (const node of dupNodes) {
-                    if (cached.nodeSummaries[node.id]) {
-                        node.duplicationSummary = cached.nodeSummaries[node.id];
-                    }
-                }
-                return;
-            }
-            console.log(`Duplication data changed since last summary (${shortHash}…). Regenerating...`);
-        } catch {
-            console.log('Duplication summary cache corrupt. Regenerating...');
-        }
-    } else if (force) {
-        console.log(`Forced duplication summary regeneration (${shortHash}…)...`);
-    } else {
-        console.log(`No duplication summary cache found (${shortHash}…). Generating...`);
-    }
-
-    const pairDescriptions = dupEdges.map(e => {
-        const d = e.duplication!;
-        const snippets = d.fragments.slice(0, 3).map((f, i) =>
-            `  Fragment ${i + 1} (${f.lines} lines, ${e.from}:${f.firstStart}-${f.firstEnd} / ${e.to}:${f.secondStart}-${f.secondEnd}):\n    ${f.fragment.slice(0, 400)}`
-        ).join('\n');
-        return `PAIR: ${e.from} <-> ${e.to} (${d.cloneCount} clones, ${d.totalLines} total duplicated lines)\n${snippets}`;
-    }).join('\n\n');
-
-    const nodeDescriptions = dupNodes.map(n => {
-        const d = n.duplication!;
-        return `${n.id} (${n.type}): ${Math.round(d.ratio * 100)}% duplicated (${d.clonedLines}/${d.totalLines} lines)`;
-    }).join('\n');
-
-    const prompt = `You are analyzing code duplication in the Lightdash backend (an open-source BI tool built with TypeScript/Express).
-
-Below are duplicated code pairs between backend nodes (controllers, services, models, clients) and per-node duplication ratios.
-
-DUPLICATED PAIRS WITH CODE FRAGMENTS:
-
-${pairDescriptions}
-
-PER-NODE DUPLICATION RATIOS:
-
-${nodeDescriptions}
-
-YOUR TASK: Provide two types of output:
-
-1. For each PAIR: Write ONE actionable refactoring suggestion (max 80 chars). Be specific about:
-   - WHAT the duplicated code does (e.g., "request validation boilerplate", "SCIM response formatting")
-   - WHERE to extract it (e.g., "base controller", "shared middleware", "utility function")
-   - HOW it helps (e.g., "single source of truth for SCIM serialization")
-
-   Good examples:
-   - "Extract SCIM list-response builder into shared scimHelpers.ts"
-   - "Shared query-result-to-response middleware would eliminate 46 lines"
-   - "DRY: extract dashboard/chart permission check into base controller"
-
-   Bad examples (NEVER do these):
-   - "Consider refactoring" (vague, no specifics)
-   - "26 lines duplicated between these files" (restating data)
-   - "These share common code" (obvious from context)
-
-2. For each NODE with >10% duplication: Write ONE line (max 70 chars) about the overall duplication pattern.
-   - Focus on the pattern: "mostly boilerplate responses" vs "business logic clones"
-   - Identify if it's structural (framework patterns) vs concerning (copied logic)
-
-   Good examples:
-   - "Boilerplate-heavy — TSOA migration would eliminate most clones"
-   - "Real logic duplication — shared query builder needed"
-   - "Router response pattern — not worth extracting"
-
-Return as JSON with "pairAdvice" (keyed by "nodeA\\0nodeB" with names alphabetically sorted) and "nodeSummaries" (keyed by node name).`;
-
-    const jsonSchema = JSON.stringify({
-        type: 'object',
-        properties: {
-            pairAdvice: {
-                type: 'object',
-                additionalProperties: { type: 'string' },
-            },
-            nodeSummaries: {
-                type: 'object',
-                additionalProperties: { type: 'string' },
-            },
-        },
-        required: ['pairAdvice', 'nodeSummaries'],
-        additionalProperties: false,
-    });
-
-    console.log(`Calling Claude to analyze ${dupEdges.length} duplication pairs, ${dupNodes.length} nodes...`);
-    const result = execSync(
-        `echo ${escapeShellArg(prompt)} | claude -p --dangerously-skip-permissions --output-format json --json-schema ${escapeShellArg(jsonSchema)} --model sonnet`,
-        { encoding: 'utf-8', maxBuffer: 10 * 1024 * 1024 },
-    );
-
-    const parsed = JSON.parse(result);
-    const output = parsed.structured_output ?? parsed;
-    const pairAdvice: Record<string, string> = output.pairAdvice;
-    const nodeSummaries: Record<string, string> = output.nodeSummaries;
-
-    for (const edge of dupEdges) {
-        const sorted = [edge.from, edge.to].sort();
-        const advice = pairAdvice[sorted.join('\0')] || pairAdvice[sorted.join(' ')];
-        if (advice) {
-            edge.duplication!.advice = advice;
-        }
-    }
-    for (const node of dupNodes) {
-        if (nodeSummaries[node.id]) {
-            node.duplicationSummary = nodeSummaries[node.id];
-        }
-    }
-
-    console.log(`Got advice for ${Object.keys(pairAdvice).length} pairs, ${Object.keys(nodeSummaries).length} nodes.`);
-
-    const cacheData = {
-        _hash: hash,
-        _generatedAt: new Date().toISOString(),
-        pairAdvice,
-        nodeSummaries,
-    };
-    fs.writeFileSync(DUPLICATION_SUMMARY_CACHE, JSON.stringify(cacheData, null, 2) + '\n');
-    console.log(`Cached to ${DUPLICATION_SUMMARY_CACHE}`);
-}
-
-// ---------------------------------------------------------------------------
-// Sentry production analytics
-// ---------------------------------------------------------------------------
-
-const SENTRY_CACHE = path.join(__dirname, '.dep-graph-sentry.json');
-const SENTRY_CACHE_TTL = 24 * 60 * 60 * 1000;
-
-interface ControllerRoute {
-    base: string;
-    subPaths: string[];
-}
-
-function parseControllerRoutes(): Record<string, ControllerRoute> {
-    const routes: Record<string, ControllerRoute> = {};
-
-    const patterns = [
-        { dir: CONTROLLERS_DIR, prefix: '' },
-        { dir: path.join(CONTROLLERS_DIR, 'v2'), prefix: 'v2/' },
-        { dir: EE_CONTROLLERS_DIR, prefix: 'ee/' },
-    ];
-
-    for (const { dir, prefix } of patterns) {
-        if (!fs.existsSync(dir)) continue;
-        const files = fs.readdirSync(dir).filter((f) => f.endsWith('.ts'));
-        for (const file of files) {
-            if (file === 'baseController.ts' || file === 'index.ts' || file.includes('.test.') || file.includes('.spec.'))
-                continue;
-            const fpath = path.join(dir, file);
-            const content = fs.readFileSync(fpath, 'utf-8');
-            const routeMatch = content.match(/@Route\(['"`]([^'"`]+)['"`]\)/);
-            if (routeMatch) {
-                const name = file.replace('.ts', '');
-                const displayName = prefix + name;
-                const subPaths = matchAll(content, /@(?:Get|Post|Put|Patch|Delete)\(['"`]([^'"`]+)['"`]\)/g);
-                routes[displayName] = { base: routeMatch[1], subPaths };
-            }
-        }
-    }
-
-    return routes;
-}
-
-function parseRouterMountPaths(): Record<string, string> {
-    const mounts: Record<string, string> = {};
-    const apiRouterPath = path.join(ROUTERS_DIR, 'apiV1Router.ts');
-    if (!fs.existsSync(apiRouterPath)) return mounts;
-
-    const content = fs.readFileSync(apiRouterPath, 'utf-8');
-    const mountRe = /apiV1Router\.use\(\s*'([^']+)'\s*,\s*(\w+)\s*\)/g;
-    let m: RegExpExecArray | null;
-    while ((m = mountRe.exec(content)) !== null) {
-        const mountPath = `/api/v1${m[1]}`;
-        const varName = m[2];
-        const files = fs.existsSync(ROUTERS_DIR) ? fs.readdirSync(ROUTERS_DIR).filter((f) => f.endsWith('.ts')) : [];
-        for (const file of files) {
-            const name = file.replace('.ts', '');
-            if (name === varName || name + 'Router' === varName || varName === name) {
-                mounts[name] = mountPath;
-                break;
-            }
-        }
-        if (!mounts[varName] && !Object.values(mounts).includes(mountPath)) {
-            const simpleName = varName.replace(/Router$/, '');
-            for (const file of files) {
-                const name = file.replace('.ts', '');
-                if (name.toLowerCase().includes(simpleName.toLowerCase())) {
-                    mounts[name] = mountPath;
-                    break;
-                }
-            }
-        }
-    }
-    return mounts;
-}
-
-interface SentryRawData {
-    transactions: Array<{ transaction: string; count: number; p95: number }>;
-    errors: Array<{ culprit: string; count: number }>;
-    spans: Array<{ transaction: string; count: number }>;
-    errorTitles: Array<{ title: string; transaction: string; count: number; issueId: string | null; groupId: string | null }>;
-}
-
-function fetchSentryData(token: string): SentryRawData {
-    const baseUrl = 'https://sentry.io/api/0/organizations/lightdash/events/';
-    const project = '5959292';
-
-    function sentryGet(params: string): any {
-        const url = `${baseUrl}?project=${project}&statsPeriod=30d&${params}`;
-        try {
-            const result = execSync(
-                `curl -sS -H "Authorization: Bearer ${token}" '${url}'`,
-                { encoding: 'utf-8', maxBuffer: 10 * 1024 * 1024, timeout: 30000 },
-            );
-            return JSON.parse(result);
-        } catch (err: any) {
-            console.warn(`Warning: Sentry API call failed: ${err.message}`);
-            return { data: [] };
-        }
-    }
-
-    console.log('Fetching Sentry data...');
-
-    const allTxnRows: any[] = [];
-    for (let page = 0; page < 10; page++) {
-        const cursor = page === 0 ? '' : `&cursor=0:${page * 100}:0`;
-        const result = sentryGet(
-            `dataset=metricsEnhanced&field=transaction&field=count()&field=p95(transaction.duration)&sort=-count()&per_page=100${cursor}`,
-        );
-        const rows = result.data || [];
-        if (rows.length === 0) break;
-        allTxnRows.push(...rows);
-    }
-
-    const routeMap = new Map<string, { count: number; p95: number }>();
-    for (const row of allTxnRows) {
-        const rawTxn: string = row.transaction;
-        if (!rawTxn.includes('/api/')) continue;
-        const route = rawTxn.replace(/^(GET|POST|PUT|PATCH|DELETE|HEAD|OPTIONS)\s+/, '');
-        const existing = routeMap.get(route) || { count: 0, p95: 0 };
-        existing.count += row['count()'] || 0;
-        existing.p95 = Math.max(existing.p95, row['p95(transaction.duration)'] || 0);
-        routeMap.set(route, existing);
-    }
-    const transactions = Array.from(routeMap.entries()).map(([route, data]) => ({
-        transaction: route,
-        count: data.count,
-        p95: data.p95,
-    }));
-
-    const errResult = sentryGet(
-        'dataset=errors&field=transaction&field=count()&sort=-count()&per_page=100',
-    );
-    const errors = (errResult.data || []).map((row: any) => ({
-        culprit: row.transaction,
-        count: row['count()'] || 0,
-    }));
-
-    const spanResult = sentryGet(
-        'dataset=metricsEnhanced&field=transaction&field=count()&query=transaction%3A*Service.*&sort=-count()&per_page=100',
-    );
-    const spans = (spanResult.data || []).map((row: any) => ({
-        transaction: row.transaction,
-        count: row['count()'] || 0,
-    }));
-
-    const errorTitleResult = sentryGet(
-        'dataset=errors&field=title&field=transaction&field=issue&field=issue.id&field=count()&sort=-count()&per_page=100',
-    );
-    const errorTitles = (errorTitleResult.data || [])
-        .filter((row: any) => row.transaction && row.title)
-        .map((row: any) => ({
-            title: row.title as string,
-            transaction: row.transaction as string,
-            count: (row['count()'] || 0) as number,
-            issueId: (row.issue as string) || null,
-            groupId: (row['issue.id'] as string) || null,
-        }));
-
-    console.log(`  ${transactions.length} transactions, ${errors.length} error culprits, ${errorTitles.length} error titles, ${spans.length} service spans.`);
-
-    return { transactions, errors, spans, errorTitles };
-}
-
-function matchSentryToNodes(
-    rawData: SentryRawData,
-    graph: GraphData,
-    controllerRoutes: Record<string, ControllerRoute>,
-    routerMounts: Record<string, string>,
-): void {
-    const routeMatchers: Array<{ nodeId: string; prefix: string; regex: RegExp }> = [];
-
-    for (const [nodeId, ctrl] of Object.entries(controllerRoutes)) {
-        if (ctrl.subPaths.length > 0) {
-            for (const sub of ctrl.subPaths) {
-                const fullPath = ctrl.base + (sub.startsWith('/') ? '' : '/') + sub;
-                const regexStr = fullPath.replace(/\{[^}]+\}/g, '[^/]+').replace(/:[^/]+/g, '[^/]+');
-                routeMatchers.push({
-                    nodeId,
-                    prefix: fullPath,
-                    regex: new RegExp(`^${regexStr}`),
-                });
-            }
-        } else {
-            const regexStr = ctrl.base.replace(/\{[^}]+\}/g, '[^/]+').replace(/:[^/]+/g, '[^/]+');
-            routeMatchers.push({
-                nodeId,
-                prefix: ctrl.base,
-                regex: new RegExp(`^${regexStr}`),
-            });
-        }
-    }
-
-    for (const [nodeId, mount] of Object.entries(routerMounts)) {
-        const regexStr = mount.replace(/\{[^}]+\}/g, '[^/]+').replace(/:[^/]+/g, '[^/]+');
-        routeMatchers.push({
-            nodeId,
-            prefix: mount,
-            regex: new RegExp(`^${regexStr}`),
-        });
-    }
-
-    routeMatchers.sort((a, b) => b.prefix.length - a.prefix.length);
-
-    const nodeEndpoints = new Map<string, Map<string, { count: number; p95: number; errors: number }>>();
-
-    const errorsByRoute = new Map<string, number>();
-    for (const err of rawData.errors) {
-        const routeMatch = err.culprit.match(/((?:GET|POST|PUT|PATCH|DELETE|HEAD|OPTIONS)\s+)?(\/api\/\S+)/);
-        if (routeMatch) {
-            const route = routeMatch[2];
-            errorsByRoute.set(route, (errorsByRoute.get(route) || 0) + err.count);
-        }
-    }
-
-    for (const txn of rawData.transactions) {
-        const route = txn.transaction.replace(/^(GET|POST|PUT|PATCH|DELETE|HEAD|OPTIONS)\s+/, '');
-        for (const matcher of routeMatchers) {
-            if (matcher.regex.test(route)) {
-                if (!nodeEndpoints.has(matcher.nodeId)) nodeEndpoints.set(matcher.nodeId, new Map());
-                const endpoints = nodeEndpoints.get(matcher.nodeId)!;
-                const existing = endpoints.get(txn.transaction) || { count: 0, p95: 0, errors: 0 };
-                existing.count += txn.count;
-                existing.p95 = Math.max(existing.p95, txn.p95);
-                let matchedErrors = 0;
-                for (const [errRoute, errCount] of errorsByRoute) {
-                    if (matcher.regex.test(errRoute)) {
-                        matchedErrors += errCount;
-                    }
-                }
-                existing.errors += matchedErrors;
-                endpoints.set(txn.transaction, existing);
-                break;
-            }
-        }
-    }
-
-    const serviceSpans = new Map<string, Array<{ name: string; count: number }>>();
-    for (const span of rawData.spans) {
-        const dotIdx = span.transaction.indexOf('.');
-        if (dotIdx < 0) continue;
-        const serviceName = span.transaction.slice(0, dotIdx);
-        const methodName = span.transaction.slice(dotIdx + 1);
-        if (!serviceSpans.has(serviceName)) serviceSpans.set(serviceName, []);
-        serviceSpans.get(serviceName)!.push({ name: methodName, count: span.count });
-    }
-
-    const nodeTopErrors = new Map<string, { title: string; count: number; issueId: string | null; groupId: string | null; transaction: string }>();
-    for (const et of rawData.errorTitles) {
-        const route = et.transaction.replace(/^(GET|POST|PUT|PATCH|DELETE|HEAD|OPTIONS)\s+/, '');
-        if (!route.includes('/api/')) continue;
-        for (const matcher of routeMatchers) {
-            if (matcher.regex.test(route)) {
-                const existing = nodeTopErrors.get(matcher.nodeId);
-                if (!existing || et.count > existing.count) {
-                    nodeTopErrors.set(matcher.nodeId, { title: et.title, count: et.count, issueId: et.issueId, groupId: et.groupId, transaction: et.transaction });
-                }
-                break;
-            }
-        }
-    }
-
-    let matchedCount = 0;
-    for (const node of graph.nodes) {
-        if (node.type === 'controller' || node.type === 'router') {
-            const endpoints = nodeEndpoints.get(node.id);
-            if (endpoints && endpoints.size > 0) {
-                const endpointList: SentryEndpoint[] = [];
-                let totalRequests = 0;
-                let totalErrors = 0;
-                let maxP95 = 0;
-
-                for (const [route, data] of endpoints) {
-                    endpointList.push({ route, count: data.count, p95Ms: data.p95, errorCount: data.errors });
-                    totalRequests += data.count;
-                    totalErrors += data.errors;
-                    maxP95 = Math.max(maxP95, data.p95);
-                }
-
-                endpointList.sort((a, b) => b.count - a.count);
-
-                const topErr = nodeTopErrors.get(node.id);
-                node.sentryActivity = {
-                    totalRequests,
-                    totalErrors,
-                    errorRate: totalRequests > 0 ? totalErrors / totalRequests : 0,
-                    maxP95Ms: maxP95,
-                    endpoints: endpointList.slice(0, 5),
-                    spans: [],
-                    topError: topErr?.title ?? null,
-                    topErrorCount: topErr?.count ?? 0,
-                    topErrorIssueId: topErr?.issueId ?? null,
-                    topErrorGroupId: topErr?.groupId ?? null,
-                    topErrorTransaction: topErr?.transaction ?? null,
-                };
-                matchedCount++;
-            }
-        } else if (node.type === 'service') {
-            const spans = serviceSpans.get(node.id);
-            if (spans && spans.length > 0) {
-                spans.sort((a, b) => b.count - a.count);
-                const totalRequests = spans.reduce((sum, s) => sum + s.count, 0);
-                node.sentryActivity = {
-                    totalRequests,
-                    totalErrors: 0,
-                    errorRate: 0,
-                    maxP95Ms: 0,
-                    endpoints: [],
-                    spans: spans.slice(0, 5),
-                    topError: null,
-                    topErrorCount: 0,
-                    topErrorIssueId: null,
-                    topErrorGroupId: null,
-                    topErrorTransaction: null,
-                };
-                matchedCount++;
-            }
-        }
-    }
-
-    console.log(`Matched Sentry data to ${matchedCount} nodes.`);
-}
-
-function collectSentryActivity(graph: GraphData, forceRefresh: boolean): void {
-    const token = process.env.SENTRY_AUTH_TOKEN;
-    if (!token) {
-        console.warn('Warning: SENTRY_AUTH_TOKEN not set. Skipping Sentry data. Set it to include production traffic metrics.');
-        return;
-    }
-
-    if (!forceRefresh && fs.existsSync(SENTRY_CACHE)) {
-        try {
-            const cached = JSON.parse(fs.readFileSync(SENTRY_CACHE, 'utf-8'));
-            const age = Date.now() - new Date(cached._generatedAt).getTime();
-            if (age < SENTRY_CACHE_TTL) {
-                const hours = Math.round(age / (60 * 60 * 1000));
-                console.log(`Sentry cache is ${hours}h old (TTL: 24h). Use --refresh to re-fetch.`);
-                for (const node of graph.nodes) {
-                    if (cached.nodeData[node.id]) {
-                        node.sentryActivity = cached.nodeData[node.id];
-                    }
-                }
-                return;
-            }
-            console.log('Sentry cache expired (>24h). Re-fetching...');
-        } catch {
-            console.log('Sentry cache corrupt. Re-fetching...');
-        }
-    }
-
-    try {
-        const controllerRoutes = parseControllerRoutes();
-        const routerMounts = parseRouterMountPaths();
-        const rawData = fetchSentryData(token);
-        matchSentryToNodes(rawData, graph, controllerRoutes, routerMounts);
-
-        const nodeData: Record<string, SentryActivity> = {};
-        for (const node of graph.nodes) {
-            if (node.sentryActivity) {
-                nodeData[node.id] = node.sentryActivity;
-            }
-        }
-
-        const cacheData = {
-            _generatedAt: new Date().toISOString(),
-            nodeData,
-        };
-        fs.writeFileSync(SENTRY_CACHE, JSON.stringify(cacheData, null, 2) + '\n');
-        console.log(`Cached to ${SENTRY_CACHE}`);
-    } catch (err: any) {
-        console.warn(`Warning: Sentry data collection failed: ${err.message}. Graph will generate without traffic data.`);
-    }
-}
-
-// ---------------------------------------------------------------------------
-// HTML template
-// ---------------------------------------------------------------------------
-
-function generateHtml(data: GraphData): string {
-    const { controllers, routers, services, models, clients, totalEdges } = data.stats;
-    const json = JSON.stringify(data);
-
-    return `<!DOCTYPE html>
-<html>
-<head>
-<meta charset="utf-8">
-<title>Lightdash Backend — Dependency Graph</title>
-<style>
-* { margin: 0; padding: 0; box-sizing: border-box; }
-body { font-family: -apple-system, BlinkMacSystemFont, system-ui, sans-serif; background: #0d1117; color: #c9d1d9; overflow: hidden; }
-
-#controls {
-  position: fixed; top: 0; left: 0; right: 0; z-index: 10;
-  background: #161b22ee; backdrop-filter: blur(8px);
-  border-bottom: 1px solid #21262d;
-  padding: 10px 16px; display: flex; gap: 8px; align-items: center; flex-wrap: wrap;
-}
-#controls input[type=text] {
-  background: #0d1117; border: 1px solid #30363d; color: #c9d1d9;
-  padding: 6px 12px; border-radius: 6px; font-size: 13px; width: 220px;
-}
-#controls input:focus { outline: none; border-color: #58a6ff; }
-.btn {
-  background: #21262d; border: 1px solid #30363d; color: #8b949e;
-  padding: 5px 12px; border-radius: 6px; font-size: 12px; cursor: pointer;
-  transition: all 0.15s;
-}
-.btn:hover { background: #30363d; color: #c9d1d9; }
-.btn.active { border-color: #58a6ff; color: #58a6ff; background: #1f6feb22; }
-.sep { width: 1px; height: 20px; background: #30363d; }
-#legend {
-  position: fixed; bottom: 16px; right: 16px; z-index: 10;
-  background: #161b22dd; backdrop-filter: blur(8px);
-  border: 1px solid #21262d; border-radius: 8px;
-  padding: 12px 16px; display: flex; flex-direction: column; gap: 6px;
-}
-.stat { font-size: 12px; color: #8b949e; }
-.stat b { font-weight: 600; }
-.stat-c b { color: #79c0ff; }
-.stat-s b { color: #7ee787; }
-.stat-m b { color: #ffa657; }
-.stat-r b { color: #d2a8ff; }
-.stat-k b { color: #f778ba; }
-
-#tooltip {
-  position: fixed; background: #1c2128; border: 1px solid #30363d;
-  border-radius: 8px; padding: 12px 16px; font-size: 12px;
-  pointer-events: none; display: none; z-index: 20; max-width: 380px;
-  box-shadow: 0 8px 24px rgba(0,0,0,0.5); line-height: 1.5;
-}
-#tooltip h3 { font-size: 14px; margin-bottom: 4px; }
-#tooltip .t-section { color: #8b949e; margin-top: 6px; font-size: 11px; text-transform: uppercase; font-weight: 600; }
-#tooltip .t-item { color: #c9d1d9; padding-left: 10px; }
-
-svg { width: 100vw; height: 100vh; }
-
-.link { stroke-opacity: 0.12; stroke-width: 1; }
-.link-uses_service { stroke: #79c0ff; }
-.link-router_uses_service { stroke: #d2a8ff; }
-.link-injects_model { stroke: #ffa657; }
-.link-injects_client { stroke: #f778ba; }
-.link-injects_service { stroke: #7ee787; stroke-dasharray: 4,3; }
-.link-shares_code { stroke: #f47067; stroke-dasharray: 4,2; stroke-width: 1.5; }
-.link-shares_code.highlighted { stroke-opacity: 0.9; stroke-width: 2.5; }
-.link.highlighted { stroke-opacity: 0.85; stroke-width: 2; }
-.link.dimmed { stroke-opacity: 0.02; }
-
-.node circle { stroke-width: 1.5; transition: opacity 0.2s; }
-.node text { pointer-events: none; transition: opacity 0.2s; }
-.node.dimmed circle, .node.dimmed text { opacity: 0.06; }
-
-.domain-hull { pointer-events: none; transition: fill-opacity 0.3s, stroke-opacity 0.3s; }
-.domain-label { pointer-events: none; font-size: 12px; font-weight: 700; text-anchor: middle; transition: fill-opacity 0.3s, font-size 0.3s; }
-.domain-only .domain-hull { fill-opacity: 0.12 !important; stroke-opacity: 0.4 !important; stroke-width: 1.5 !important; }
-.domain-only .domain-label { fill-opacity: 0.9 !important; font-size: 14px !important; }
-
-#heatmap-legend {
-  display: none; flex-direction: column; gap: 4px; margin-top: 8px;
-  padding-top: 8px; border-top: 1px solid #21262d;
-}
-.heatmap-bar { width: 100%; height: 10px; border-radius: 4px; background: linear-gradient(to right, #3fb950, #d29922, #f47067); }
-.heatmap-labels { display: flex; justify-content: space-between; font-size: 10px; color: #8b949e; }
-
-#guide {
-  position: fixed; top: 52px; left: 16px; z-index: 15;
-  background: #161b22ee; backdrop-filter: blur(8px);
-  border: 1px solid #21262d; border-radius: 10px;
-  width: 320px; max-height: calc(100vh - 76px);
-  overflow-y: auto; padding: 0;
-  transform: translateX(-340px);
-  transition: transform 0.25s ease;
-  scrollbar-width: thin; scrollbar-color: #30363d transparent;
-}
-#guide.open { transform: translateX(0); }
-#guide::-webkit-scrollbar { width: 6px; }
-#guide::-webkit-scrollbar-thumb { background: #30363d; border-radius: 3px; }
-#guide::-webkit-scrollbar-track { background: transparent; }
-#guide .guide-header {
-  position: sticky; top: 0; z-index: 1;
-  background: #161b22; border-bottom: 1px solid #21262d;
-  padding: 12px 16px; display: flex; justify-content: space-between; align-items: center;
-}
-#guide .guide-header h2 { font-size: 14px; font-weight: 600; color: #c9d1d9; margin: 0; }
-#guide .guide-close {
-  background: none; border: none; color: #8b949e; cursor: pointer;
-  font-size: 18px; line-height: 1; padding: 0 4px;
-}
-#guide .guide-close:hover { color: #c9d1d9; }
-#guide .guide-body { padding: 12px 16px 16px; }
-.guide-section { margin-bottom: 12px; }
-.guide-section:last-child { margin-bottom: 0; }
-.guide-toggle {
-  display: flex; align-items: center; gap: 6px; width: 100%;
-  background: none; border: none; color: #c9d1d9; cursor: pointer;
-  font-size: 13px; font-weight: 600; padding: 6px 0; text-align: left;
-}
-.guide-toggle:hover { color: #58a6ff; }
-.guide-toggle .chevron {
-  display: inline-block; font-size: 10px; color: #8b949e;
-  transition: transform 0.15s; width: 12px; text-align: center;
-}
-.guide-toggle[aria-expanded="true"] .chevron { transform: rotate(90deg); }
-.guide-content {
-  display: none; padding: 4px 0 4px 18px;
-  font-size: 12px; color: #8b949e; line-height: 1.6;
-}
-.guide-content.open { display: block; }
-.guide-content p { margin: 0 0 6px; }
-.guide-content p:last-child { margin: 0; }
-.guide-content kbd {
-  display: inline-block; background: #21262d; border: 1px solid #30363d;
-  border-radius: 3px; padding: 0 5px; font-size: 11px; font-family: inherit;
-  color: #c9d1d9; line-height: 1.6;
-}
-.guide-content .color-dot {
-  display: inline-block; width: 8px; height: 8px; border-radius: 50%;
-  vertical-align: middle; margin-right: 4px;
-}
-#detail {
-  position: fixed; top: 52px; right: 16px; width: 380px; max-height: calc(100vh - 68px);
-  background: #161b22; border: 1px solid #21262d; border-radius: 8px;
-  z-index: 15; overflow: hidden; display: flex; flex-direction: column;
-  transition: transform 0.25s ease, opacity 0.25s ease;
-  transform: translateX(400px); opacity: 0; pointer-events: none;
-}
-#detail.open { transform: translateX(0); opacity: 1; pointer-events: auto; }
-#detail .detail-header {
-  position: sticky; top: 0; z-index: 1;
-  background: #161b22; border-bottom: 1px solid #21262d;
-  padding: 12px 16px; display: flex; justify-content: space-between; align-items: center;
-  flex-shrink: 0;
-}
-#detail .detail-header h2 { font-size: 14px; font-weight: 600; color: #c9d1d9; margin: 0; word-break: break-all; }
-#detail .detail-close {
-  background: none; border: none; color: #8b949e; cursor: pointer;
-  font-size: 18px; line-height: 1; padding: 0 4px; flex-shrink: 0;
-}
-#detail .detail-close:hover { color: #c9d1d9; }
-#detail .detail-body {
-  padding: 12px 16px 16px; overflow-y: auto; flex: 1;
-}
-#detail .detail-body::-webkit-scrollbar { width: 6px; }
-#detail .detail-body::-webkit-scrollbar-thumb { background: #30363d; border-radius: 3px; }
-#detail .detail-body::-webkit-scrollbar-track { background: transparent; }
-#detail .detail-body a { color: #58a6ff; text-decoration: none; }
-#detail .detail-body a:hover { text-decoration: underline; }
-#detail .detail-body .t-section { font-size: 11px; text-transform: uppercase; color: #484f58; margin-top: 8px; letter-spacing: 0.05em; }
-#detail .detail-body .t-item { font-size: 12px; color: #c9d1d9; padding-left: 10px; margin-top: 3px; font-family: 'SF Mono', 'Fira Code', monospace; }
-</style>
-</head>
-<body>
-<div id="controls">
-  <input type="text" id="search" placeholder="Search...">
-  <div class="sep"></div>
-  <button class="btn active" data-filter="all">All</button>
-  <button class="btn" data-filter="controller">Controllers</button>
-  <button class="btn" data-filter="router">Routers</button>
-  <button class="btn" data-filter="service">Services</button>
-  <button class="btn" data-filter="model">Models</button>
-  <button class="btn" data-filter="client">Clients</button>
-  <div class="sep"></div>
-  <label class="stat" style="display:flex;align-items:center;gap:4px;" title="How many hops from the selected node to highlight">Depth <input type="number" id="depth" value="2" min="1" max="10" style="width:42px;background:#0d1117;border:1px solid #30363d;color:#c9d1d9;padding:3px 6px;border-radius:4px;font-size:12px;text-align:center;"></label>
-  <div class="sep"></div>
-  <label class="stat" style="display:flex;align-items:center;gap:4px;" title="What drives node size and color">Size <select id="size-mode" style="background:#0d1117;border:1px solid #30363d;color:#c9d1d9;padding:3px 6px;border-radius:4px;font-size:12px;"><option value="edges">Edges</option><option value="lines">Lines</option><option value="cyclomatic">Cyclomatic</option><option value="cognitive">Cognitive</option><option value="commits">Commits</option><option value="authors">Authors</option><option value="churn">Churn (6mo)</option><option value="health">Health</option><option value="duplication">Duplication</option><option value="traffic">Traffic</option><option value="errors">Errors</option></select></label>
-  <div class="sep"></div>
-  <button class="btn active" id="btn-arrows">Arrows</button>
-  <div class="sep"></div>
-  <label class="stat" style="display:flex;align-items:center;gap:4px;">Group <select id="group-mode" style="background:#0d1117;border:1px solid #30363d;color:#c9d1d9;padding:3px 6px;border-radius:4px;font-size:12px;"><option value="domain">Domain</option><option value="layer">Layer</option></select></label>
-  <button class="btn active" id="btn-nodes">Nodes</button>
-  <div class="sep"></div>
-  <button class="btn" id="btn-ee" style="display:none">EE</button>
-  <div class="sep" id="sep-ee" style="display:none"></div>
-  <button class="btn" id="btn-reset">Reset</button>
-  <div class="sep"></div>
-  <button class="btn" id="btn-guide" title="How to use this graph">?</button>
-</div>
-<div id="guide">
-  <div class="guide-header">
-    <h2>How to use this graph</h2>
-    <button class="guide-close" id="guide-close">&times;</button>
-  </div>
-  <div class="guide-body">
-    <div class="guide-section">
-      <button class="guide-toggle" aria-expanded="true"><span class="chevron">&#9654;</span> Navigation</button>
-      <div class="guide-content open">
-        <p><b>Pan</b> &mdash; Click and drag the background</p>
-        <p><b>Zoom</b> &mdash; Scroll wheel or pinch</p>
-        <p><b>Move a node</b> &mdash; Drag it to reposition</p>
-        <p><b>Reset view</b> &mdash; Click the <kbd>Reset</kbd> button</p>
-      </div>
-    </div>
-    <div class="guide-section">
-      <button class="guide-toggle" aria-expanded="true"><span class="chevron">&#9654;</span> Selecting &amp; exploring</button>
-      <div class="guide-content open">
-        <p><b>Click a node</b> to select it. Its dependencies are highlighted with animated edges that ripple outward.</p>
-        <p><b>Depth</b> controls how many hops away from the selected node are shown (1&ndash;10).</p>
-        <p><b>Click the background</b> to deselect.</p>
-        <p><b>Search</b> filters nodes by name &mdash; matching nodes and their direct connections stay visible.</p>
-      </div>
-    </div>
-    <div class="guide-section">
-      <button class="guide-toggle" aria-expanded="false"><span class="chevron">&#9654;</span> Node types &amp; colors</button>
-      <div class="guide-content">
-        <p><span class="color-dot" style="background:#79c0ff"></span><b>Controller</b> &mdash; HTTP/API entry points</p>
-        <p><span class="color-dot" style="background:#d2a8ff"></span><b>Router</b> &mdash; Route definitions</p>
-        <p><span class="color-dot" style="background:#7ee787"></span><b>Service</b> &mdash; Business logic layer</p>
-        <p><span class="color-dot" style="background:#ffa657"></span><b>Model</b> &mdash; Database/data access layer</p>
-        <p><span class="color-dot" style="background:#f778ba"></span><b>Client</b> &mdash; External service clients</p>
-        <p>Filter by type using the toolbar buttons.</p>
-      </div>
-    </div>
-    <div class="guide-section">
-      <button class="guide-toggle" aria-expanded="false"><span class="chevron">&#9654;</span> Edge types</button>
-      <div class="guide-content">
-        <p><span style="color:#79c0ff">&#9473;</span> <b>uses_service</b> &mdash; Controller calls a service</p>
-        <p><span style="color:#d2a8ff">&#9473;</span> <b>router_uses_service</b> &mdash; Router calls a service</p>
-        <p><span style="color:#ffa657">&#9473;</span> <b>injects_model</b> &mdash; Service depends on a model</p>
-        <p><span style="color:#f778ba">&#9473;</span> <b>injects_client</b> &mdash; Service depends on a client</p>
-        <p><span style="color:#7ee787">- -</span> <b>injects_service</b> &mdash; Service-to-service (dashed)</p>
-        <p><span style="color:#f47067">- -</span> <b>shares_code</b> &mdash; Code duplication link (dashed)</p>
-      </div>
-    </div>
-    <div class="guide-section">
-      <button class="guide-toggle" aria-expanded="false"><span class="chevron">&#9654;</span> Size modes</button>
-      <div class="guide-content">
-        <p><b>Size</b> controls node radius and automatically sets an appropriate color scheme:</p>
-        <p>&bull; <b>Edges</b> &mdash; Dependency count (colored by type)</p>
-        <p>&bull; <b>Lines</b> &mdash; Source lines of code (colored by type)</p>
-        <p>&bull; <b>Cyclomatic</b> &mdash; Number of execution paths (colored by type)</p>
-        <p>&bull; <b>Cognitive</b> &mdash; Human readability difficulty (colored by type)</p>
-        <p>&bull; <b>Commits / Authors / Churn</b> &mdash; Git activity (colored by type)</p>
-        <p>&bull; <b>Health</b> &mdash; Composite score, green&#8594;red heatmap</p>
-        <p>&bull; <b>Duplication</b> &mdash; Shared code blocks, green&#8594;red heatmap</p>
-        <p>&bull; <b>Traffic</b> &mdash; Sentry 30d requests, yellow&#8594;red heatmap (requires --sentry)</p>
-        <p>&bull; <b>Errors</b> &mdash; Sentry 30d errors, yellow&#8594;red heatmap (requires --sentry)</p>
-      </div>
-    </div>
-    <div class="guide-section">
-      <button class="guide-toggle" aria-expanded="false"><span class="chevron">&#9654;</span> Domains &amp; toggles</button>
-      <div class="guide-content">
-        <p><b>Domains</b> groups nodes by their domain directory, drawing a hull boundary around each group. Toggle off for a flat layout.</p>
-        <p><b>Nodes</b> toggles node and edge visibility. Turn off to see only the domain regions &mdash; useful for a high-level architectural overview.</p>
-        <p><b>Arrows</b> shows dependency direction on highlighted edges.</p>
-        <p><b>EE</b> (if visible) toggles enterprise-edition nodes.</p>
-      </div>
-    </div>
-    <div class="guide-section">
-      <button class="guide-toggle" aria-expanded="false"><span class="chevron">&#9654;</span> Tooltips &amp; health score</button>
-      <div class="guide-content">
-        <p><b>Hover any node</b> to see its metrics: edges, lines of code, complexity scores, git churn, recent commits, and duplication info.</p>
-        <p>The <b>health score</b> (0&ndash;100) is a composite: 30% coupling, 25% complexity, 30% churn, 15% commit frequency. Lower is better.</p>
-      </div>
-    </div>
-  </div>
-</div>
-<div id="detail">
-  <div class="detail-header">
-    <h2 id="detail-title"></h2>
-    <button class="detail-close" id="detail-close">&times;</button>
-  </div>
-  <div class="detail-body" id="detail-body"></div>
-</div>
-<div id="legend">
-  <span class="stat stat-c"><b>${controllers}</b> controllers</span>
-  <span class="stat stat-r"><b>${routers}</b> routers</span>
-  <span class="stat stat-s"><b>${services}</b> services</span>
-  <span class="stat stat-m"><b>${models}</b> models</span>
-  <span class="stat stat-k"><b>${clients}</b> clients</span>
-  <span class="stat"><b>${totalEdges}</b> edges</span>
-  <div id="heatmap-legend">
-    <div class="heatmap-labels"><span>Healthy</span><span>Hot</span></div>
-    <div class="heatmap-bar"></div>
-  </div>
-</div>
-<div id="tooltip"></div>
-<svg></svg>
-
-<script src="https://d3js.org/d3.v7.min.js"></script>
-<script>
-const raw = ${json};
-
-const color = { controller: '#79c0ff', router: '#d2a8ff', service: '#7ee787', model: '#ffa657', client: '#f778ba' };
-const colorDark = { controller: '#1f6feb', router: '#8957e5', service: '#238636', model: '#9e6a03', client: '#da3633' };
-
-const W = window.innerWidth, H = window.innerHeight;
-
-const nodeIdx = {};
-raw.nodes.forEach((n, i) => { nodeIdx[n.id] = i; });
-
-const connCount = {};
-raw.edges.forEach(e => {
-  connCount[e.from] = (connCount[e.from] || 0) + 1;
-  connCount[e.to] = (connCount[e.to] || 0) + 1;
-});
-
-function calcRadius(n) {
-  const mode = document.getElementById('size-mode').value;
-  switch (mode) {
-    case 'lines': return Math.max(4, 3 + Math.sqrt((n.lineCount || 100) / 30) * 2.2);
-    case 'cyclomatic': return Math.max(4, 3 + Math.sqrt((n.complexity?.cyclomatic || 1) / 5) * 2.2);
-    case 'cognitive': return Math.max(4, 3 + Math.sqrt((n.complexity?.cognitive || 1) / 5) * 2.2);
-    case 'commits': return Math.max(4, 3 + Math.sqrt((n.gitActivity?.commits || 1) / 8) * 2.2);
-    case 'authors': return Math.max(4, 3 + Math.sqrt(n.gitActivity?.authors || 1) * 3);
-    case 'churn': return Math.max(4, 3 + Math.sqrt((n.gitActivity?.churn || 0) / 2) * 2.5);
-    case 'health': return Math.max(4, 3 + (n.healthScore || 0) * 16);
-    case 'duplication': return Math.max(4, 3 + Math.sqrt((n.duplication?.ratio || 0) * 100) * 2);
-    case 'traffic': return Math.max(4, 3 + Math.log10((n.sentryActivity?.totalRequests || 1)) * 2.5);
-    case 'errors': return Math.max(4, 3 + Math.log10((n.sentryActivity?.totalErrors || 1)) * 2.5);
-    default: return Math.max(4, 3 + Math.sqrt(connCount[n.id] || 1) * 2.2);
-  }
-}
-
-const nodes = raw.nodes.map(n => ({ ...n, r: calcRadius(n) }));
-
-// Health score: normalized composite of coupling, complexity, churn, commits
-const hMetrics = {
-  coupling: raw.nodes.map(n => connCount[n.id] || 0),
-  complexity: raw.nodes.map(n => n.complexity?.cyclomatic || 0),
-  churn: raw.nodes.map(n => n.gitActivity?.churn || 0),
-  commits: raw.nodes.map(n => n.gitActivity?.commits || 0),
-};
-function maxNorm(vals) { const mx = Math.max(...vals); return mx === 0 ? vals.map(() => 0) : vals.map(v => v / mx); }
-const nCoup = maxNorm(hMetrics.coupling), nComp = maxNorm(hMetrics.complexity);
-const nChurn = maxNorm(hMetrics.churn), nComm = maxNorm(hMetrics.commits);
-const healthColorScale = d3.scaleLinear().domain([0, 0.5, 1]).range(['#3fb950', '#d29922', '#f47067']).interpolate(d3.interpolateRgb);
-const healthStrokeScale = d3.scaleLinear().domain([0, 0.5, 1]).range(['#238636', '#9e6a03', '#da3633']).interpolate(d3.interpolateRgb);
-nodes.forEach((n, i) => { n.healthScore = nCoup[i] * 0.3 + nComp[i] * 0.25 + nChurn[i] * 0.3 + nComm[i] * 0.15; });
-
-const trafficMax = Math.max(...raw.nodes.map(n => Math.log10((n.sentryActivity?.totalRequests || 0) + 1)));
-function trafficNorm(n) { if (!n.sentryActivity || trafficMax === 0) return -1; return Math.log10(n.sentryActivity.totalRequests + 1) / trafficMax; }
-const trafficColorScale = d3.scaleLinear().domain([0, 0.5, 1]).range(['#e2b340', '#e8853a', '#f47067']).interpolate(d3.interpolateRgb);
-const trafficStrokeScale = d3.scaleLinear().domain([0, 0.5, 1]).range(['#9e6a03', '#b85c1e', '#da3633']).interpolate(d3.interpolateRgb);
-
-const errorsMax = Math.max(...raw.nodes.map(n => Math.log10((n.sentryActivity?.totalErrors || 0) + 1)));
-function errorsNorm(n) { if (!n.sentryActivity || errorsMax === 0) return -1; return Math.log10((n.sentryActivity.totalErrors || 0) + 1) / errorsMax; }
-const errorsColorScale = d3.scaleLinear().domain([0, 0.5, 1]).range(['#d29922', '#f47067', '#da3633']).interpolate(d3.interpolateRgb);
-const errorsStrokeScale = d3.scaleLinear().domain([0, 0.5, 1]).range(['#9e6a03', '#da3633', '#8b1a1a']).interpolate(d3.interpolateRgb);
-
-function getColorMode() { const m = document.getElementById('size-mode').value; if (m === 'health' || m === 'duplication' || m === 'traffic' || m === 'errors') return m; return 'type'; }
-function getNodeColor(d) { const m = getColorMode(); if (m === 'health') return healthColorScale(d.healthScore); if (m === 'duplication') return healthColorScale(d.duplication?.ratio || 0); if (m === 'traffic') { const t = trafficNorm(d); return t < 0 ? '#30363d' : trafficColorScale(t); } if (m === 'errors') { const e = errorsNorm(d); return e < 0 ? '#30363d' : e === 0 ? '#3fb950' : errorsColorScale(e); } return color[d.type]; }
-function getNodeStroke(d) { const m = getColorMode(); if (m === 'health') return healthStrokeScale(d.healthScore); if (m === 'duplication') return healthStrokeScale(d.duplication?.ratio || 0); if (m === 'traffic') { const t = trafficNorm(d); return t < 0 ? '#21262d' : trafficStrokeScale(t); } if (m === 'errors') { const e = errorsNorm(d); return e < 0 ? '#21262d' : e === 0 ? '#238636' : errorsStrokeScale(e); } return colorDark[d.type]; }
-
-const links = raw.edges.map(e => ({
-  source: nodeIdx[e.from],
-  target: nodeIdx[e.to],
-  type: e.type
-}));
-
-const outgoing = {}, incoming = {};
-raw.edges.forEach(e => {
-  if (!outgoing[e.from]) outgoing[e.from] = [];
-  outgoing[e.from].push({ id: e.to, type: e.type });
-  if (!incoming[e.to]) incoming[e.to] = [];
-  incoming[e.to].push({ id: e.from, type: e.type });
-});
-
-const svg = d3.select('svg').attr('width', W).attr('height', H);
-
-// Arrowhead markers (5 edge types × dim/bright)
-const defs = svg.append('defs');
-const edgeTypes = [
-  { type: 'uses_service', color: '#79c0ff' },
-  { type: 'router_uses_service', color: '#d2a8ff' },
-  { type: 'injects_model', color: '#ffa657' },
-  { type: 'injects_client', color: '#f778ba' },
-  { type: 'injects_service', color: '#7ee787' },
-  { type: 'shares_code', color: '#f47067' },
-];
-edgeTypes.forEach(({ type, color: c }) => {
-  [{ suffix: '', opacity: 0.15, w: 8, h: 6 }, { suffix: 'hi-', opacity: 0.9, w: 5, h: 3.75 }].forEach(({ suffix, opacity, w, h }) => {
-    defs.append('marker')
-      .attr('id', 'arrow-' + suffix + type)
-      .attr('viewBox', '0 0 10 6')
-      .attr('refX', 10).attr('refY', 3)
-      .attr('markerWidth', w).attr('markerHeight', h)
-      .attr('orient', 'auto')
-      .append('path')
-      .attr('d', 'M0,0 L10,3 L0,6Z')
-      .attr('fill', c)
-      .attr('fill-opacity', opacity);
-  });
-});
-
-const g = svg.append('g');
-const zoomBehavior = d3.zoom().scaleExtent([0.05, 10]).on('zoom', e => g.attr('transform', e.transform));
-svg.call(zoomBehavior);
-
-const colX = { controller: W * 0.08, router: W * 0.20, service: W * 0.40, model: W * 0.68, client: W * 0.88 };
-
-let domainNames = [];
-let domainColor = {};
-let domainCenters = {};
-const typeOffsetX = { controller: -50, router: -25, service: 0, model: 25, client: 50 };
-
-const domainMap = {};
-nodes.forEach(n => {
-  if (n.domain) {
-    if (!domainMap[n.domain]) domainMap[n.domain] = [];
-    domainMap[n.domain].push(n);
-  }
-});
-domainNames = Object.keys(domainMap).sort();
-const dCols = Math.ceil(Math.sqrt(domainNames.length * (W / H)));
-const dRows = Math.ceil(domainNames.length / dCols);
-const cellW = W / (dCols + 1);
-const cellH = (H - 60) / (dRows + 1);
-const palette = d3.schemeTableau10;
-
-domainNames.forEach((name, i) => {
-  const col = i % dCols;
-  const row = Math.floor(i / dCols);
-  const cx = cellW * (col + 1);
-  const cy = cellH * (row + 1) + 40;
-  domainCenters[name] = { x: cx, y: cy };
-  domainColor[name] = palette[i % palette.length];
-  domainMap[name].forEach(n => {
-    n.domainX = cx + (typeOffsetX[n.type] || 0);
-    n.domainY = cy;
-  });
-});
-
-let groupMode = 'domain';
-
-const layerTypes = ['controller', 'router', 'service', 'model', 'client'];
-const layerLabels = { controller: 'Controllers', router: 'Routers', service: 'Services', model: 'Models', client: 'Clients' };
-let layerCenters = {};
-const layerMap = {};
-nodes.forEach(n => {
-  if (!layerMap[n.type]) layerMap[n.type] = [];
-  layerMap[n.type].push(n);
-});
-const lCols = Math.ceil(Math.sqrt(layerTypes.length * (W / H)));
-const lRows = Math.ceil(layerTypes.length / lCols);
-const lCellW = W / (lCols + 1);
-const lCellH = (H - 60) / (lRows + 1);
-layerTypes.forEach((type, i) => {
-  const col = i % lCols;
-  const row = Math.floor(i / lCols);
-  const cx = lCellW * (col + 1);
-  const cy = lCellH * (row + 1) + 40;
-  layerCenters[type] = { x: cx, y: cy };
-  (layerMap[type] || []).forEach(n => {
-    n.layerX = cx;
-    n.layerY = cy;
-  });
-});
-
-const sim = d3.forceSimulation(nodes)
-  .force('link', d3.forceLink(links).distance(60).strength(0.1))
-  .force('charge', d3.forceManyBody().strength(-150).distanceMax(300))
-  .force('collision', d3.forceCollide().radius(d => d.r + 18).strength(0.9).iterations(2))
-  .force('x', d3.forceX(d => d.domainX || W/2).strength(0.3))
-  .force('y', d3.forceY(d => d.domainY || H/2).strength(0.3));
-
-const colLabels = g.append('g').attr('class', 'col-labels').style('display', 'none');
-['Controllers', 'Routers', 'Services', 'Models', 'Clients'].forEach((label, i) => {
-  const types = ['controller', 'router', 'service', 'model', 'client'];
-  colLabels.append('text')
-    .attr('x', colX[types[i]])
-    .attr('y', 30)
-    .attr('text-anchor', 'middle')
-    .attr('fill', color[types[i]])
-    .attr('font-size', '13px')
-    .attr('font-weight', '600')
-    .attr('opacity', 0.4)
-    .text(label);
-});
-
-function applyGroupLayout() {
-  if (groupMode === 'domain') {
-    sim.force('link', d3.forceLink(links).distance(60).strength(0.1))
-      .force('charge', d3.forceManyBody().strength(-150).distanceMax(300))
-      .force('collision', d3.forceCollide().radius(d => d.r + 18).strength(0.9).iterations(2))
-      .force('center', null)
-      .force('x', d3.forceX(d => d.domainX || W/2).strength(0.3))
-      .force('y', d3.forceY(d => d.domainY || H/2).strength(0.3));
-    colLabels.style('display', 'none');
-  } else {
-    sim.force('link', d3.forceLink(links).distance(60).strength(0.1))
-      .force('charge', d3.forceManyBody().strength(-150).distanceMax(300))
-      .force('collision', d3.forceCollide().radius(d => d.r + 18).strength(0.9).iterations(2))
-      .force('center', null)
-      .force('x', d3.forceX(d => d.layerX || W/2).strength(0.3))
-      .force('y', d3.forceY(d => d.layerY || H/2).strength(0.3));
-    colLabels.style('display', 'none');
-  }
-  sim.alpha(0.8).restart();
-}
-
-const hullLayer = g.insert('g', ':first-child').attr('class', 'hull-layer');
-
-const link = g.append('g').selectAll('line').data(links).join('line')
-  .attr('class', d => 'link link-' + d.type)
-  .attr('marker-end', d => 'url(#arrow-' + d.type + ')');
-
-let dragged = false;
-const node = g.append('g').selectAll('g').data(nodes).join('g')
-  .attr('class', 'node')
-  .call(d3.drag()
-    .on('start', (e, d) => { dragged = false; d.fx = d.x; d.fy = d.y; })
-    .on('drag', (e, d) => { dragged = true; if (!e.active) sim.alphaTarget(0.3).restart(); d.fx = e.x; d.fy = e.y; })
-    .on('end', (e, d) => { if (!e.active) sim.alphaTarget(0); if (!dragged) { d.fx = null; d.fy = null; } }));
-
-node.append('circle')
-  .attr('r', d => d.r)
-  .attr('fill', d => getNodeColor(d))
-  .attr('stroke', d => getNodeStroke(d))
-  .attr('stroke-width', 1.5)
-  .attr('stroke-dasharray', d => d.ee ? '3,2' : null);
-
-node.append('text')
-  .text(d => d.id)
-  .attr('dx', d => d.r + 3)
-  .attr('dy', 3)
-  .attr('fill', d => color[d.type])
-  .attr('font-size', d => d.r > 8 ? '10px' : '8px')
-  .attr('opacity', 0.7);
-
-const tip = d3.select('#tooltip');
-
-function fmtNum(n) { if (n >= 1e6) return (n / 1e6).toFixed(1) + 'M'; if (n >= 1e3) return (n / 1e3).toFixed(1) + 'K'; return String(n); }
-function fmtMs(ms) { if (ms >= 1000) return (ms / 1000).toFixed(1) + 's'; return Math.round(ms) + 'ms'; }
-
-function buildNodeContent(d, opts) {
-  const interactive = opts && opts.interactive;
-  let h = '';
-  if (!interactive) {
-    h += '<h3 style="color:' + color[d.type] + '">' + d.id + '</h3>';
-  }
-  h += '<span style="color:#8b949e">' + d.type + '</span>';
-  if (d.domain) h += ' <span style="color:#6e7681">· ' + d.domain + '</span>';
-  const pct = Math.round((1 - d.healthScore) * 100);
-  h += '<div class="t-section">health score</div>';
-  h += '<div class="t-item"><span style="color:' + healthColorScale(d.healthScore) + ';font-weight:700;font-size:18px">' + pct + '</span><span style="color:#484f58;font-size:12px"> / 100</span></div>';
-  h += '<div class="t-item" style="font-size:11px;color:#8b949e;margin-top:2px">';
-  h += (connCount[d.id]||0) + ' edges · ' + (d.lineCount||0) + ' lines · ' + (d.gitActivity?.churn||0) + ' churn · ' + (d.gitActivity?.commits||0) + ' commits';
-  h += '</div>';
-  if (d.complexity) {
-    h += '<div class="t-item" style="font-size:11px;color:#8b949e;margin-top:2px">Cyclomatic: <b style="color:#c9d1d9">' + d.complexity.cyclomatic + '</b> · Cognitive: <b style="color:#c9d1d9">' + d.complexity.cognitive + '</b> · Max/fn: <b style="color:#c9d1d9">' + d.complexity.maxFunctionCyclomatic + '</b></div>';
-  }
-  if (d.healthSummary) {
-    h += '<div style="color:' + healthColorScale(d.healthScore) + ';font-style:italic;padding-left:10px;margin-top:4px;font-size:11px">' + d.healthSummary + '</div>';
-  }
-  if (d.duplication && d.duplication.ratio > 0) {
-    const dupPct = Math.round(d.duplication.ratio * 100);
-    const dupEdges = raw.edges.filter(e => e.type === 'shares_code' && (e.from === d.id || e.to === d.id));
-    h += '<div class="t-section">duplication</div>';
-    h += '<div class="t-item"><span style="color:#f47067;font-weight:700">' + dupPct + '%</span> duplicated <span style="color:#484f58">(' + d.duplication.clonedLines + ' / ' + d.duplication.totalLines + ' lines)</span>';
-    if (dupEdges.length > 0) h += ' · <span style="color:#484f58">' + dupEdges.length + ' shared</span>';
-    h += '</div>';
-    if (d.duplicationSummary) {
-      h += '<div style="color:#f47067;font-style:italic;padding-left:10px;margin-top:4px;font-size:11px">' + d.duplicationSummary + '</div>';
-    }
-    dupEdges.forEach(e => {
-      const other = e.from === d.id ? e.to : e.from;
-      const dup = e.duplication;
-      if (dup) {
-        h += '<div class="t-item" style="font-size:11px;margin-top:4px"><span style="color:#f47067">↔</span> <b style="color:#c9d1d9">' + other + '</b> <span style="color:#484f58">' + dup.cloneCount + ' clone' + (dup.cloneCount > 1 ? 's' : '') + ', ' + dup.totalLines + ' lines</span></div>';
-        if (dup.advice) {
-          h += '<div style="color:#d29922;font-style:italic;padding-left:18px;margin-top:1px;font-size:11px">' + dup.advice + '</div>';
-        }
-      }
-    });
-  }
-  if (d.sentryActivity) {
-    const sa = d.sentryActivity;
-    h += '<div class="t-section">sentry (30d)</div>';
-    h += '<div class="t-item"><span style="color:#58a6ff;font-weight:700">' + fmtNum(sa.totalRequests) + '</span> requests';
-    if (sa.maxP95Ms > 0) h += ' · p95 <span style="color:#c9d1d9;font-weight:600">' + fmtMs(sa.maxP95Ms) + '</span>';
-    if (sa.totalErrors > 0) h += ' · <span style="color:#f47067">' + fmtNum(sa.totalErrors) + ' errors (' + (sa.errorRate * 100).toFixed(2) + '%)</span>';
-    h += '</div>';
-    if (sa.topError) {
-      const shortErr = sa.topError.length > 60 ? sa.topError.slice(0, 57) + '…' : sa.topError;
-      if (interactive && sa.topErrorGroupId) {
-        var issueUrl = 'https://lightdash.sentry.io/issues/' + sa.topErrorGroupId + '/events/?project=5959292';
-        if (sa.topErrorTransaction) issueUrl += '&query=' + encodeURIComponent('transaction:"' + sa.topErrorTransaction + '"');
-        h += '<div class="t-item" style="font-size:11px;margin-top:2px"><span style="color:#f47067">⚡</span> <a href="' + issueUrl + '" target="_blank" style="color:#f0883e">' + shortErr + '</a> <span style="color:#484f58">(' + fmtNum(sa.topErrorCount) + ')</span></div>';
-      } else if (interactive) {
-        h += '<div class="t-item" style="font-size:11px;margin-top:2px"><span style="color:#f47067">⚡</span> <span style="color:#f0883e">' + shortErr + '</span> <span style="color:#484f58">(' + fmtNum(sa.topErrorCount) + ')</span></div>';
-      } else {
-        h += '<div class="t-item" style="font-size:11px;margin-top:2px"><span style="color:#f47067">⚡</span> <span style="color:#f0883e">' + shortErr + '</span> <span style="color:#484f58">(' + fmtNum(sa.topErrorCount) + ')</span></div>';
-      }
-    }
-    if (sa.endpoints && sa.endpoints.length > 0) {
-      h += '<div class="t-section" style="margin-top:4px">top endpoints</div>';
-      sa.endpoints.slice(0, 3).forEach(ep => {
-        const shortRoute = ep.route.length > 45 ? '...' + ep.route.slice(-42) : ep.route;
-        h += '<div class="t-item" style="font-size:11px"><span style="color:#58a6ff">' + fmtNum(ep.count).padStart(5) + '</span>  ' + shortRoute;
-        if (ep.p95Ms > 0) h += '  <span style="color:#484f58">p95 ' + fmtMs(ep.p95Ms) + '</span>';
-        if (ep.errorCount > 0) h += '  <span style="color:#f47067">' + fmtNum(ep.errorCount) + ' err</span>';
-        h += '</div>';
-      });
-    }
-    if (sa.spans && sa.spans.length > 0) {
-      h += '<div class="t-section" style="margin-top:4px">service spans</div>';
-      sa.spans.slice(0, 3).forEach(sp => {
-        h += '<div class="t-item" style="font-size:11px"><span style="color:#58a6ff">' + fmtNum(sp.count).padStart(5) + '</span>  ' + sp.name + '</div>';
-      });
-    }
-  }
-  if (d.gitActivity) {
-    h += '<div class="t-section">git activity</div>';
-    h += '<div class="t-item">' + d.gitActivity.commits + ' commits · ' + d.gitActivity.authors + ' authors · ' + d.gitActivity.churn + ' recent (6mo)</div>';
-    if (d.gitSummary) {
-      h += '<div style="color:#7ee787;font-style:italic;padding-left:10px;margin-top:4px;font-size:11px;white-space:pre-line">' + d.gitSummary + '</div>';
-    }
-    if (interactive || !d.gitSummary) {
-      if (d.gitActivity.recentCommits && d.gitActivity.recentCommits.length > 0) {
-        h += '<div class="t-section" style="margin-top:6px">recent commits</div>';
-        d.gitActivity.recentCommits.forEach(c => {
-          const msg = c.message.length > 50 ? c.message.slice(0, 50) + '…' : c.message;
-          if (interactive) {
-            h += '<div class="t-item" style="font-size:11px"><a href="https://github.com/lightdash/lightdash/commit/' + c.hash + '" target="_blank" style="color:#484f58">' + c.hash + '</a>  ' + msg + '  <span style="color:#484f58">' + c.relativeDate + '</span></div>';
-          } else {
-            h += '<div class="t-item" style="font-size:11px"><span style="color:#484f58">' + c.hash + '</span>  ' + msg + '  <span style="color:#484f58">' + c.relativeDate + '</span></div>';
-          }
-        });
-      }
-    }
-  }
-  return h;
-}
-
-node.on('mouseover', (ev, d) => {
-  if (selected === d.id) return;
-  tip.html(buildNodeContent(d, { interactive: false })).style('display', 'block');
-})
-.on('mousemove', ev => {
-  const x = Math.min(ev.clientX + 16, W - 400);
-  tip.style('left', x + 'px').style('top', (ev.clientY - 10) + 'px');
-})
-.on('mouseout', () => tip.style('display', 'none'));
-
-let selected = null;
-node.on('click', (ev, d) => {
-  ev.stopPropagation();
-  tip.style('display', 'none');
-  if (selected === d.id) { selected = null; clearHL(); closeDetail(); return; }
-  selected = d.id;
-  highlightNode(d);
-  openDetail(d);
-});
-svg.on('click', () => {
-  selected = null; clearHL(); closeDetail();
-  document.querySelectorAll('[data-filter]').forEach(b => b.classList.remove('active'));
-  document.querySelector('[data-filter="all"]').classList.add('active');
-});
-
-document.getElementById('depth').addEventListener('input', () => {
-  if (selected) {
-    const d = nodes.find(n => n.id === selected);
-    if (d) highlightNode(d);
-  }
-});
-
-function updateHeatmapLegend() {
-  const mode = getColorMode();
-  const isHeatmap = mode !== 'type';
-  document.getElementById('heatmap-legend').style.display = isHeatmap ? 'flex' : 'none';
-  if (!isHeatmap) return;
-  const labels = document.querySelectorAll('.heatmap-labels span');
-  const bar = document.querySelector('.heatmap-bar');
-  if (mode === 'duplication') { labels[0].textContent = 'No duplication'; labels[1].textContent = 'High duplication'; bar.style.background = 'linear-gradient(to right, #3fb950, #d29922, #f47067)'; }
-  else if (mode === 'traffic') { labels[0].textContent = 'Low traffic'; labels[1].textContent = 'High traffic'; bar.style.background = 'linear-gradient(to right, #e2b340, #e8853a, #f47067)'; }
-  else if (mode === 'errors') { labels[0].textContent = 'No errors'; labels[1].textContent = 'High errors'; bar.style.background = 'linear-gradient(to right, #d29922, #f47067, #da3633)'; }
-  else { labels[0].textContent = 'Healthy'; labels[1].textContent = 'Hot'; bar.style.background = 'linear-gradient(to right, #3fb950, #d29922, #f47067)'; }
-}
-
-document.getElementById('size-mode').addEventListener('change', () => {
-  nodes.forEach(n => { n.r = calcRadius(n); });
-  node.select('circle').attr('r', d => d.r);
-  node.select('text').attr('dx', d => d.r + 3).attr('font-size', d => d.r > 8 ? '10px' : '8px');
-  sim.force('collision', d3.forceCollide().radius(d => d.r + 18).strength(0.9).iterations(2));
-  sim.alpha(0.3).restart();
-  node.select('circle')
-    .transition().duration(400)
-    .attr('fill', d => getNodeColor(d))
-    .attr('stroke', d => getNodeStroke(d));
-  updateHeatmapLegend();
-});
-
-function getConnected(id, depth) {
-  const downMap = new Map([[id, 0]]);
-  let downFrontier = new Set([id]);
-  for (let d = 0; d < depth; d++) {
-    const next = new Set();
-    links.forEach(l => {
-      const src = typeof l.source === 'object' ? l.source.id : nodes[l.source].id;
-      const tgt = typeof l.target === 'object' ? l.target.id : nodes[l.target].id;
-      if (downFrontier.has(src) && !downMap.has(tgt)) { downMap.set(tgt, d + 1); next.add(tgt); }
-    });
-    if (next.size === 0) break;
-    downFrontier = next;
-  }
-
-  const upMap = new Map([[id, 0]]);
-  let upFrontier = new Set([id]);
-  for (let d = 0; d < depth; d++) {
-    const next = new Set();
-    links.forEach(l => {
-      const src = typeof l.source === 'object' ? l.source.id : nodes[l.source].id;
-      const tgt = typeof l.target === 'object' ? l.target.id : nodes[l.target].id;
-      if (upFrontier.has(tgt) && !upMap.has(src)) { upMap.set(src, d + 1); next.add(src); }
-    });
-    if (next.size === 0) break;
-    upFrontier = next;
-  }
-
-  const merged = new Map(downMap);
-  upMap.forEach((dep, nid) => {
-    if (!merged.has(nid) || dep < merged.get(nid)) merged.set(nid, dep);
-  });
-  return merged;
-}
-
-function getDepth() {
-  return Math.max(1, parseInt(document.getElementById('depth').value) || 1);
-}
-
-function highlightNode(d) {
-  link.interrupt();
-  const depth = getDepth();
-  const conn = getConnected(d.id, depth);
-  node.classed('dimmed', n => !conn.has(n.id));
-  link.classed('dimmed', l => {
-    const s = typeof l.source === 'object' ? l.source.id : nodes[l.source].id;
-    const t = typeof l.target === 'object' ? l.target.id : nodes[l.target].id;
-    return !(conn.has(s) && conn.has(t));
-  });
-  link.classed('highlighted', l => {
-    const s = typeof l.source === 'object' ? l.source.id : nodes[l.source].id;
-    const t = typeof l.target === 'object' ? l.target.id : nodes[l.target].id;
-    return conn.has(s) && conn.has(t);
-  });
-  updateMarkers();
-
-  const edgesBySource = {};
-  link.filter('.highlighted').each(function(l) {
-    const src = typeof l.source === 'object' ? l.source.id : nodes[l.source].id;
-    const tgt = typeof l.target === 'object' ? l.target.id : nodes[l.target].id;
-    const srcDepth = conn.get(src) ?? Infinity;
-    const tgtDepth = conn.get(tgt) ?? Infinity;
-    const outward = srcDepth <= tgtDepth;
-    const origin = outward ? src : tgt;
-    if (!edgesBySource[origin]) edgesBySource[origin] = [];
-    edgesBySource[origin].push({ el: d3.select(this), l, outward });
-    d3.select(this).style('stroke-opacity', 0).attr('marker-end', null);
-  });
-
-  const animated = new Set();
-  function animateFrom(nodeId) {
-    const edges = edgesBySource[nodeId];
-    if (!edges) return;
-    edges.forEach(e => {
-      const el = e.el;
-      if (animated.has(el.node())) return;
-      animated.add(el.node());
-      const x1 = +el.attr('x1'), y1 = +el.attr('y1');
-      const x2 = +el.attr('x2'), y2 = +el.attr('y2');
-      const len = Math.hypot(x2 - x1, y2 - y1);
-      if (len === 0) return;
-      const src = typeof e.l.source === 'object' ? e.l.source.id : nodes[e.l.source].id;
-      const tgt = typeof e.l.target === 'object' ? e.l.target.id : nodes[e.l.target].id;
-      const destId = e.outward ? tgt : src;
-      el.style('stroke-opacity', null)
-        .attr('stroke-dasharray', len)
-        .attr('stroke-dashoffset', e.outward ? len : -len);
-      el.transition()
-        .duration(400 + Math.random() * 200)
-        .ease(d3.easeCubicOut)
-        .attr('stroke-dashoffset', 0)
-        .on('end', function() {
-          const self = d3.select(this);
-          if (e.l.type === 'injects_service') self.attr('stroke-dasharray', '4,3');
-          else if (e.l.type === 'shares_code') self.attr('stroke-dasharray', '4,2');
-          else self.attr('stroke-dasharray', null);
-          self.attr('stroke-dashoffset', null);
-          if (arrowsEnabled && e.l.type !== 'shares_code') self.attr('marker-end', 'url(#arrow-hi-' + e.l.type + ')');
-          animateFrom(destId);
-        });
-    });
-  }
-  animateFrom(d.id);
-}
-
-function clearHL() {
-  link.interrupt();
-  node.classed('dimmed', false);
-  link.classed('dimmed', false).classed('highlighted', false);
-  link.each(function(l) {
-    const el = d3.select(this);
-    el.style('stroke-opacity', null);
-    if (l.type === 'injects_service') el.attr('stroke-dasharray', '4,3');
-    else if (l.type === 'shares_code') el.attr('stroke-dasharray', '4,2');
-    else el.attr('stroke-dasharray', null);
-    el.attr('stroke-dashoffset', null);
-  });
-  updateMarkers();
-}
-
-let arrowsEnabled = true;
-function updateMarkers() {
-  link.attr('marker-end', function(l) {
-    if (!arrowsEnabled) return null;
-    if (l.type === 'shares_code') return null;
-    if (d3.select(this).classed('dimmed')) return null;
-    const isHL = d3.select(this).classed('highlighted');
-    return 'url(#arrow-' + (isHL ? 'hi-' : '') + l.type + ')';
-  });
-}
-
-document.getElementById('btn-arrows').addEventListener('click', () => {
-  arrowsEnabled = !arrowsEnabled;
-  document.getElementById('btn-arrows').classList.toggle('active', arrowsEnabled);
-  updateMarkers();
-});
-
-document.getElementById('group-mode').addEventListener('change', e => {
-  groupMode = e.target.value;
-  applyGroupLayout();
-});
-
-let nodesVisible = true;
-function applyNodeVisibility() {
-  node.style('opacity', nodesVisible ? null : 0).style('pointer-events', nodesVisible ? null : 'none');
-  link.style('opacity', nodesVisible ? null : 0);
-  svg.node().classList.toggle('domain-only', !nodesVisible);
-}
-document.getElementById('btn-nodes').addEventListener('click', () => {
-  nodesVisible = !nodesVisible;
-  document.getElementById('btn-nodes').classList.toggle('active', nodesVisible);
-  applyNodeVisibility();
-});
-
-document.getElementById('search').addEventListener('input', e => {
-  const q = e.target.value.toLowerCase();
-  if (!q) { clearHL(); return; }
-  const matches = new Set();
-  nodes.forEach(n => { if (n.id.toLowerCase().includes(q)) matches.add(n.id); });
-  const conn = new Set(matches);
-  links.forEach(l => {
-    const s = typeof l.source === 'object' ? l.source.id : nodes[l.source].id;
-    const t = typeof l.target === 'object' ? l.target.id : nodes[l.target].id;
-    if (matches.has(s) || matches.has(t)) { conn.add(s); conn.add(t); }
-  });
-  node.classed('dimmed', n => !conn.has(n.id));
-  link.classed('dimmed', l => {
-    const s = typeof l.source === 'object' ? l.source.id : nodes[l.source].id;
-    const t = typeof l.target === 'object' ? l.target.id : nodes[l.target].id;
-    return !(matches.has(s) || matches.has(t));
-  });
-  link.classed('highlighted', l => {
-    const s = typeof l.source === 'object' ? l.source.id : nodes[l.source].id;
-    const t = typeof l.target === 'object' ? l.target.id : nodes[l.target].id;
-    return matches.has(s) || matches.has(t);
-  });
-  updateMarkers();
-});
-
-document.querySelectorAll('[data-filter]').forEach(btn => {
-  btn.addEventListener('click', () => {
-    document.querySelectorAll('[data-filter]').forEach(b => b.classList.remove('active'));
-    btn.classList.add('active');
-    const type = btn.dataset.filter;
-    if (type === 'all') { clearHL(); return; }
-    const matches = new Set();
-    nodes.forEach(n => { if (n.type === type) matches.add(n.id); });
-    node.classed('dimmed', n => !matches.has(n.id));
-    link.classed('dimmed', l => {
-      const s = typeof l.source === 'object' ? l.source.id : nodes[l.source].id;
-      const t = typeof l.target === 'object' ? l.target.id : nodes[l.target].id;
-      return !(matches.has(s) && matches.has(t));
-    });
-    updateMarkers();
-  });
-});
-
-document.getElementById('btn-reset').addEventListener('click', () => {
-  selected = null; clearHL(); closeDetail();
-  document.getElementById('search').value = '';
-  document.querySelectorAll('[data-filter]').forEach(b => b.classList.remove('active'));
-  document.querySelector('[data-filter="all"]').classList.add('active');
-  document.getElementById('size-mode').value = 'edges';
-  groupMode = 'domain';
-  document.getElementById('group-mode').value = 'domain';
-  nodes.forEach(n => { n.r = calcRadius(n); });
-  node.select('circle').attr('r', d => d.r).attr('fill', d => getNodeColor(d)).attr('stroke', d => getNodeStroke(d));
-  node.select('text').attr('dx', d => d.r + 3).attr('font-size', d => d.r > 8 ? '10px' : '8px');
-  updateHeatmapLegend();
-  if (!nodesVisible) {
-    nodesVisible = true;
-    document.getElementById('btn-nodes').classList.add('active');
-    applyNodeVisibility();
-  }
-  applyGroupLayout();
-  svg.transition().duration(500).call(zoomBehavior.transform, d3.zoomIdentity);
-});
-
-// Guide sidebar
-const guideEl = document.getElementById('guide');
-document.getElementById('btn-guide').addEventListener('click', () => {
-  guideEl.classList.toggle('open');
-  document.getElementById('btn-guide').classList.toggle('active', guideEl.classList.contains('open'));
-});
-document.getElementById('guide-close').addEventListener('click', () => {
-  guideEl.classList.remove('open');
-  document.getElementById('btn-guide').classList.remove('active');
-});
-document.querySelectorAll('.guide-toggle').forEach(btn => {
-  btn.addEventListener('click', () => {
-    const expanded = btn.getAttribute('aria-expanded') === 'true';
-    btn.setAttribute('aria-expanded', String(!expanded));
-    btn.nextElementSibling.classList.toggle('open', !expanded);
-  });
-});
-
-// Detail sidebar
-const detailEl = document.getElementById('detail');
-const detailTitle = document.getElementById('detail-title');
-const detailBody = document.getElementById('detail-body');
-function openDetail(d) {
-  detailTitle.textContent = d.id;
-  detailTitle.style.color = color[d.type];
-  detailBody.innerHTML = buildNodeContent(d, { interactive: true });
-  detailEl.classList.add('open');
-}
-function closeDetail() {
-  detailEl.classList.remove('open');
-}
-document.getElementById('detail-close').addEventListener('click', () => {
-  closeDetail();
-  selected = null;
-  clearHL();
-});
-
-// EE toggle
-const hasEe = nodes.some(n => n.ee);
-if (hasEe) {
-  document.getElementById('btn-ee').style.display = '';
-  document.getElementById('sep-ee').style.display = '';
-  let eeVisible = true;
-  const eeBtn = document.getElementById('btn-ee');
-  eeBtn.classList.add('active');
-  function applyEeVisibility() {
-    node.each(function(d) {
-      if (d.ee) {
-        d3.select(this).style('display', eeVisible ? null : 'none');
-      }
-    });
-    link.each(function(l) {
-      const s = typeof l.source === 'object' ? l.source.id : nodes[l.source].id;
-      const t = typeof l.target === 'object' ? l.target.id : nodes[l.target].id;
-      const sNode = nodes.find(n => n.id === s);
-      const tNode = nodes.find(n => n.id === t);
-      if ((sNode && sNode.ee) || (tNode && tNode.ee)) {
-        d3.select(this).style('display', eeVisible ? null : 'none');
-      }
-    });
-  }
-  applyEeVisibility();
-  eeBtn.addEventListener('click', () => {
-    eeVisible = !eeVisible;
-    eeBtn.classList.toggle('active', eeVisible);
-    applyEeVisibility();
-  });
-}
-
-function padHull(points, pad) {
-  if (points.length < 3) return points;
-  const cx = d3.mean(points, p => p[0]);
-  const cy = d3.mean(points, p => p[1]);
-  return points.map(p => {
-    const dx = p[0] - cx, dy = p[1] - cy;
-    const len = Math.sqrt(dx * dx + dy * dy) || 1;
-    return [p[0] + dx / len * pad, p[1] + dy / len * pad];
-  });
-}
-
-function smoothHull(points) {
-  if (points.length < 3) return '';
-  const n = points.length;
-  let d = 'M' + points[0].join(',');
-  for (let i = 0; i < n; i++) {
-    const p0 = points[i];
-    const p1 = points[(i + 1) % n];
-    const p2 = points[(i + 2) % n];
-    const mx1 = (p0[0] + p1[0]) / 2, my1 = (p0[1] + p1[1]) / 2;
-    const mx2 = (p1[0] + p2[0]) / 2, my2 = (p1[1] + p2[1]) / 2;
-    if (i === 0) d = 'M' + mx1 + ',' + my1;
-    d += ' Q' + p1[0] + ',' + p1[1] + ' ' + mx2 + ',' + my2;
-  }
-  return d + 'Z';
-}
-
-sim.on('tick', () => {
-  link.each(function(d) {
-    const dx = d.target.x - d.source.x;
-    const dy = d.target.y - d.source.y;
-    const len = Math.sqrt(dx * dx + dy * dy) || 1;
-    const r = d.target.r || 6;
-    const el = d3.select(this);
-    el.attr('x1', d.source.x).attr('y1', d.source.y)
-      .attr('x2', d.target.x - dx / len * r)
-      .attr('y2', d.target.y - dy / len * r);
-  });
-  node.attr('transform', d => \`translate(\${d.x},\${d.y})\`);
-
-  hullLayer.selectAll('*').remove();
-  if (groupMode === 'domain') {
-    domainNames.forEach(name => {
-      const pts = nodes.filter(n => n.domain === name).map(n => [n.x, n.y]);
-      if (pts.length < 2) return;
-      if (pts.length === 2) {
-        const [a, b] = pts;
-        const dx = b[0] - a[0], dy = b[1] - a[1];
-        const len = Math.sqrt(dx * dx + dy * dy) || 1;
-        const nx = -dy / len * 25, ny = dx / len * 25;
-        const expanded = [[a[0]+nx,a[1]+ny],[b[0]+nx,b[1]+ny],[b[0]-nx,b[1]-ny],[a[0]-nx,a[1]-ny]];
-        hullLayer.append('path')
-          .attr('class', 'domain-hull')
-          .attr('d', 'M' + expanded.map(p => p.join(',')).join('L') + 'Z')
-          .attr('fill', domainColor[name]).attr('fill-opacity', 0.04)
-          .attr('stroke', domainColor[name]).attr('stroke-opacity', 0.15)
-          .attr('stroke-width', 1);
-      } else {
-        const hull = d3.polygonHull(pts);
-        if (hull) {
-          const padded = padHull(hull, 25);
-          hullLayer.append('path')
-            .attr('class', 'domain-hull')
-            .attr('d', smoothHull(padded))
-            .attr('fill', domainColor[name]).attr('fill-opacity', 0.04)
-            .attr('stroke', domainColor[name]).attr('stroke-opacity', 0.15)
-            .attr('stroke-width', 1);
-        }
-      }
-      const cx = d3.mean(pts, p => p[0]);
-      const cy = d3.mean(pts, p => p[1]);
-      const minY = d3.min(pts, p => p[1]);
-      hullLayer.append('text')
-        .attr('class', 'domain-label')
-        .attr('x', cx).attr('y', minY - 18)
-        .attr('fill', domainColor[name]).attr('fill-opacity', 0.6)
-        .text(name);
-    });
-  } else {
-    layerTypes.forEach(type => {
-      const pts = nodes.filter(n => n.type === type).map(n => [n.x, n.y]);
-      if (pts.length < 2) return;
-      if (pts.length === 2) {
-        const [a, b] = pts;
-        const dx = b[0] - a[0], dy = b[1] - a[1];
-        const len = Math.sqrt(dx * dx + dy * dy) || 1;
-        const nx = -dy / len * 25, ny = dx / len * 25;
-        const expanded = [[a[0]+nx,a[1]+ny],[b[0]+nx,b[1]+ny],[b[0]-nx,b[1]-ny],[a[0]-nx,a[1]-ny]];
-        hullLayer.append('path')
-          .attr('class', 'domain-hull')
-          .attr('d', 'M' + expanded.map(p => p.join(',')).join('L') + 'Z')
-          .attr('fill', color[type]).attr('fill-opacity', 0.04)
-          .attr('stroke', color[type]).attr('stroke-opacity', 0.15)
-          .attr('stroke-width', 1);
-      } else {
-        const hull = d3.polygonHull(pts);
-        if (hull) {
-          const padded = padHull(hull, 25);
-          hullLayer.append('path')
-            .attr('class', 'domain-hull')
-            .attr('d', smoothHull(padded))
-            .attr('fill', color[type]).attr('fill-opacity', 0.04)
-            .attr('stroke', color[type]).attr('stroke-opacity', 0.15)
-            .attr('stroke-width', 1);
-        }
-      }
-      const cx = d3.mean(pts, p => p[0]);
-      const cy = d3.mean(pts, p => p[1]);
-      const minY = d3.min(pts, p => p[1]);
-      hullLayer.append('text')
-        .attr('class', 'domain-label')
-        .attr('x', cx).attr('y', minY - 18)
-        .attr('fill', color[type]).attr('fill-opacity', 0.6)
-        .text(layerLabels[type]);
-    });
-  }
-});
-</script>
-</body>
-</html>`;
-}
-
-// ---------------------------------------------------------------------------
-// Helpers
-// ---------------------------------------------------------------------------
-
-function matchAll(str: string, re: RegExp): string[] {
-    const results: string[] = [];
-    let m: RegExpExecArray | null;
-    while ((m = re.exec(str)) !== null) {
-        results.push(m[1]);
-    }
-    return results;
-}
-
-function unique(arr: string[]): string[] {
-    return [...new Set(arr)].sort();
-}
-
-// ---------------------------------------------------------------------------
-// Domain classification (via Claude CLI)
-// ---------------------------------------------------------------------------
-
-const DOMAIN_CACHE = path.join(__dirname, '.dep-graph-domains.json');
-
-function computeNodeHash(graph: GraphData): string {
-    const ids = graph.nodes
-        .map((n) => n.id)
-        .sort()
-        .join('\n');
-    return crypto.createHash('sha256').update(ids).digest('hex');
-}
-
-function classifyDomains(graph: GraphData, force: boolean): Record<string, string[]> {
-    const hash = computeNodeHash(graph);
-    const shortHash = hash.slice(0, 12);
-
-    // Check cache
-    if (!force && fs.existsSync(DOMAIN_CACHE)) {
-        const cached = JSON.parse(fs.readFileSync(DOMAIN_CACHE, 'utf-8'));
-        if (cached._hash === hash) {
-            const count = Object.keys(cached.domains).length;
-            console.log(`Domain cache up to date (${count} domains, ${shortHash}…). Use --refresh to reclassify.`);
-            return cached.domains;
-        }
-        console.log(`Graph changed since last classification (${shortHash}…). Reclassifying...`);
-    } else if (force) {
-        console.log(`Forced reclassification (${shortHash}…)...`);
-    } else {
-        console.log(`No domain cache found (${shortHash}…). Classifying...`);
-    }
-
-    // Build prompt
-    const nodeList = graph.nodes
-        .map((n) => `- ${n.id} (${n.type})`)
-        .join('\n');
-    const edgeList = graph.edges
-        .map((e) => `- ${e.from} -> ${e.to} (${e.type})`)
-        .join('\n');
-
-    const prompt = `You are classifying backend dependency-injection nodes into business domains for a dependency graph visualization.
-
-Below are ${graph.nodes.length} nodes from the Lightdash backend (an open-source BI tool). Each node is a controller, service, model, or client.
-
-NODES:
-${nodeList}
-
-EDGES:
-${edgeList}
-
-TASK: Group ALL nodes into 14-18 business domains based on their business function, NOT their technical layer.
-
-Rules:
-- Nodes that share a name root belong together (e.g., SpaceService, SpaceModel, spaceController -> "Spaces")
-- Every node must appear in exactly one domain
-- Domain names must be short (1-2 words)
-- Use the edges to inform grouping: tightly connected nodes likely belong together
-- Use EXACTLY the domain names listed below when they apply
-
-Expected domains (use these names, assign all matching nodes to them):
-- "Spaces" — spaceController, SpaceService, SpaceModel, SpacePermissionService, SpacePermissionModel
-- "Dashboards" — dashboardController, v2/DashboardController, DashboardService, DashboardModel
-- "Saved Charts" — savedChartController, v2/SavedChartController, SavedChartService, SavedChartModel
-- "SQL Runner" — sqlRunnerController, SavedSqlService, SavedSqlModel
-- "Explores" — exploreController, runQueryController, v2/QueryController, metricsExplorerController, funnelController, MetricsExplorerService, FunnelService, AsyncQueryService, QueryHistoryModel, PivotTableService
-- "Projects" — projectController, v2/ParametersController, ProjectService, ProjectModel, ProjectParametersService, ProjectParametersModel, CoderService, WarehouseAvailableTablesModel
-- "Organizations" — organizationController, OrganizationService, OrganizationModel, OrganizationMemberProfileModel, OrganizationAllowedEmailDomainsModel, OrganizationWarehouseCredentialsModel, UserWarehouseCredentialsModel
-- "Roles & Permissions" — OrganizationRolesController, ProjectRolesController, RolesService, RolesModel, PermissionsService, groupsController, GroupService, GroupsModel
-- "User Auth" — userController, UserService, UserModel, SessionModel, OpenIdIdentityModel, PasswordResetLinkModel, InviteLinkModel, OauthService, OauthModel, PersonalAccessTokenService, PersonalAccessTokenModel, OnboardingModel
-- "Scheduling" — schedulerController, csvController, SchedulerService, SchedulerModel, SchedulerClient, JobModel, CsvService, DownloadFileService, DownloadFileModel, DownloadAuditModel
-- "Content" — v2/ContentController, ContentService, ContentModel, pinningController, PinningService, PinnedListModel, ResourceViewItemModel, renameController, RenameService, shareController, ShareService, ShareModel, PromoteService
-- "Catalog" — catalogController, CatalogService, CatalogModel, ChangesetController, ChangesetService, ChangesetModel, TagsModel
-- "Notifications" — notificationsController, commentsController, NotificationService, NotificationsModel, CommentService, CommentModel
-- "Git Integration" — gitIntegrationController, githubController, gitlabController, GitIntegrationService, GithubAppService, GitlabAppService, GithubAppInstallationsModel, GitlabAppInstallationsModel
-- "Slack" — slackController, SlackIntegrationService, SlackService, SlackClient, SlackAuthenticationModel, UnfurlService
-- "Infrastructure" — sshController, SshKeyPairService, SshKeyPairModel, S3Client, S3CacheClient, ResultsFileStorageClient, EmailClient, EmailModel, HealthService, MigrationModel, EncryptionUtil (if present)
-
-For any remaining nodes not listed above, assign them to the closest matching domain or create a new domain if needed. Do NOT split the above domains — keep them intact.`;
-
-    const jsonSchema = JSON.stringify({
-        type: 'object',
-        properties: {
-            domains: {
-                type: 'object',
-                additionalProperties: {
-                    type: 'array',
-                    items: { type: 'string' },
-                },
-            },
-        },
-        required: ['domains'],
-        additionalProperties: false,
-    });
-
-    console.log('Calling Claude for classification...');
-    const result = execSync(
-        `echo ${escapeShellArg(prompt)} | claude -p --dangerously-skip-permissions --output-format json --json-schema ${escapeShellArg(jsonSchema)} --model sonnet`,
-        { encoding: 'utf-8', maxBuffer: 10 * 1024 * 1024 },
-    );
-
-    const parsed = JSON.parse(result);
-    const output = parsed.structured_output ?? parsed;
-    const domains: Record<string, string[]> = output.domains;
-
-    const domainCount = Object.keys(domains).length;
-    const classifiedCount = Object.values(domains).reduce(
-        (s, a) => s + a.length,
-        0,
-    );
-    console.log(
-        `Got ${domainCount} domains covering ${classifiedCount} nodes (expected ${graph.nodes.length}).`,
-    );
-
-    // Write cache
-    const cacheData = {
-        _hash: hash,
-        _generatedAt: new Date().toISOString(),
-        domains,
-    };
-    fs.writeFileSync(DOMAIN_CACHE, JSON.stringify(cacheData, null, 2) + '\n');
-    console.log(`Cached to ${DOMAIN_CACHE}`);
-
-    return domains;
-}
-
-// ---------------------------------------------------------------------------
-// Git summary (via Claude CLI)
-// ---------------------------------------------------------------------------
-
-const SUMMARY_CACHE = path.join(__dirname, '.dep-graph-summaries.json');
-
-function computeSummaryHash(graph: GraphData): string {
-    const parts = graph.nodes
-        .filter((n) => n.gitActivity && n.gitActivity.recentCommits.length > 0)
-        .map(
-            (n) =>
-                `${n.id}:${n.gitActivity!.recentCommits.map((c) => c.hash).join(',')}`,
-        )
-        .sort()
-        .join('\n');
-    return crypto.createHash('sha256').update(parts).digest('hex');
-}
-
-function summarizeGitActivity(
-    graph: GraphData,
-    force: boolean,
-): Record<string, string> {
-    const hash = computeSummaryHash(graph);
-    const shortHash = hash.slice(0, 12);
-
-    if (!force && fs.existsSync(SUMMARY_CACHE)) {
-        const cached = JSON.parse(fs.readFileSync(SUMMARY_CACHE, 'utf-8'));
-        if (cached._hash === hash) {
-            const count = Object.keys(cached.summaries).length;
-            console.log(
-                `Summary cache up to date (${count} nodes, ${shortHash}…). Use --refresh to regenerate.`,
-            );
-            return cached.summaries;
-        }
-        console.log(
-            `Git activity changed since last summary (${shortHash}…). Regenerating...`,
-        );
-    } else if (force) {
-        console.log(`Forced summary regeneration (${shortHash}…)...`);
-    } else {
-        console.log(`No summary cache found (${shortHash}…). Generating...`);
-    }
-
-    const nodesWithCommits = graph.nodes.filter(
-        (n) => n.gitActivity && n.gitActivity.recentCommits.length > 0,
-    );
-
-    if (nodesWithCommits.length === 0) {
-        console.log('No nodes with recent commits to summarize.');
-        return {};
-    }
-
-    const nodeDescriptions = nodesWithCommits
-        .map((n) => {
-            const commits = n
-                .gitActivity!.recentCommits.map(
-                    (c) => `  ${c.hash} ${c.message} (${c.author}, ${c.relativeDate})`,
-                )
-                .join('\n');
-            return `${n.id} (${n.type}):\n${commits}`;
-        })
-        .join('\n\n');
-
-    const prompt = `You are summarizing recent git activity for nodes in a backend dependency graph of Lightdash (an open-source BI tool).
-
-For each node below, write a 3-line summary of what changed recently (one line per recent commit). Each line should be max 60 chars, present tense, specific about what changed (not just "updated" or "modified"). Focus on the business intent. Separate lines with newlines.
-
-NODES AND THEIR RECENT COMMITS:
-
-${nodeDescriptions}
-
-Return a summary for each node. Use ONLY the node name as the key (e.g., "ProjectService", not "ProjectService (service)").`;
-
-    const jsonSchema = JSON.stringify({
-        type: 'object',
-        properties: {
-            summaries: {
-                type: 'object',
-                additionalProperties: { type: 'string' },
-            },
-        },
-        required: ['summaries'],
-        additionalProperties: false,
-    });
-
-    console.log(
-        `Calling Claude to summarize ${nodesWithCommits.length} nodes...`,
-    );
-    const result = execSync(
-        `echo ${escapeShellArg(prompt)} | claude -p --dangerously-skip-permissions --output-format json --json-schema ${escapeShellArg(jsonSchema)} --model sonnet`,
-        { encoding: 'utf-8', maxBuffer: 10 * 1024 * 1024 },
-    );
-
-    const parsed = JSON.parse(result);
-    const output = parsed.structured_output ?? parsed;
-    const summaries: Record<string, string> = output.summaries;
-
-    const summaryCount = Object.keys(summaries).length;
-    console.log(
-        `Got summaries for ${summaryCount} nodes (${nodesWithCommits.length} requested).`,
-    );
-
-    const cacheData = {
-        _hash: hash,
-        _generatedAt: new Date().toISOString(),
-        summaries,
-    };
-    fs.writeFileSync(
-        SUMMARY_CACHE,
-        JSON.stringify(cacheData, null, 2) + '\n',
-    );
-    console.log(`Cached to ${SUMMARY_CACHE}`);
-
-    return summaries;
-}
-
-// ---------------------------------------------------------------------------
-// Health summary (via Claude CLI)
-// ---------------------------------------------------------------------------
-
-const HEALTH_SUMMARY_CACHE = path.join(__dirname, '.dep-graph-health-summaries.json');
-
-function summarizeHealthScores(
-    graph: GraphData,
-    force: boolean,
-): Record<string, string> {
-    const edgeCounts: Record<string, number> = {};
-    graph.edges.forEach((e) => {
-        edgeCounts[e.from] = (edgeCounts[e.from] || 0) + 1;
-        edgeCounts[e.to] = (edgeCounts[e.to] || 0) + 1;
-    });
-
-    const parts = graph.nodes
-        .map((n) => {
-            const edges = edgeCounts[n.id] || 0;
-            const cyc = n.complexity?.cyclomatic || 0;
-            const cog = n.complexity?.cognitive || 0;
-            let line = `${n.id}:${edges}:${n.lineCount || 0}:${cyc}:${cog}:${n.gitActivity?.churn || 0}:${n.gitActivity?.commits || 0}`;
-            if (n.sentryActivity) {
-                line += `:${n.sentryActivity.totalRequests}:${n.sentryActivity.totalErrors}:${n.sentryActivity.topError || ''}:${n.sentryActivity.topErrorCount}`;
-            }
-            return line;
-        })
-        .sort()
-        .join('\n');
-    const hash = crypto.createHash('sha256').update(parts).digest('hex');
-    const shortHash = hash.slice(0, 12);
-
-    if (!force && fs.existsSync(HEALTH_SUMMARY_CACHE)) {
-        const cached = JSON.parse(fs.readFileSync(HEALTH_SUMMARY_CACHE, 'utf-8'));
-        if (cached._hash === hash) {
-            const count = Object.keys(cached.summaries).length;
-            console.log(
-                `Health summary cache up to date (${count} nodes, ${shortHash}…). Use --refresh to regenerate.`,
-            );
-            return cached.summaries;
-        }
-        console.log(
-            `Health metrics changed since last summary (${shortHash}…). Regenerating...`,
-        );
-    } else if (force) {
-        console.log(`Forced health summary regeneration (${shortHash}…)...`);
-    } else {
-        console.log(`No health summary cache found (${shortHash}…). Generating...`);
-    }
-
-    const nodeDescriptions = graph.nodes
-        .map((n) => {
-            const edges = edgeCounts[n.id] || 0;
-            const lines = n.lineCount || 0;
-            const cyc = n.complexity?.cyclomatic || 0;
-            const cog = n.complexity?.cognitive || 0;
-            const churn = n.gitActivity?.churn || 0;
-            const commits = n.gitActivity?.commits || 0;
-            const authors = n.gitActivity?.authors || 0;
-            let desc = `${n.id} (${n.type}): ${edges} edges, ${lines} lines, cyclomatic ${cyc}, cognitive ${cog}, ${churn} churn (6mo), ${commits} total commits, ${authors} authors`;
-            if (n.sentryActivity) {
-                const sa = n.sentryActivity;
-                const fmtReqs = sa.totalRequests >= 1e6 ? (sa.totalRequests / 1e6).toFixed(1) + 'M' : sa.totalRequests >= 1e3 ? (sa.totalRequests / 1e3).toFixed(1) + 'K' : String(sa.totalRequests);
-                desc += ` | ${fmtReqs} requests, ${(sa.errorRate * 100).toFixed(1)}% error rate`;
-                if (sa.topError) {
-                    desc += `, top error: "${sa.topError}" (${sa.topErrorCount} occurrences)`;
-                }
-            }
-            return desc;
-        })
-        .join('\n');
-
-    const prompt = `You are analyzing backend dependency-injection nodes in Lightdash (an open-source BI tool).
-
-The user already sees the raw numbers (edges, lines, cyclomatic, cognitive, churn, commits) in the tooltip. Your job is to provide INSIGHT the numbers alone don't show — what the combination of metrics means, what the risk is, or what action to take.
-
-NEVER restate the numbers. Instead explain WHY the combination matters or WHAT to do about it.
-
-For each node, write ONE line (max 70 chars). Focus on:
-- Cross-metric patterns: "high complexity + high churn = regression risk"
-- Actionable advice: "split into query-building and execution halves"
-- Risk classification: what kind of problem this node represents
-- Comparisons: "god service — owns too many concerns"
-- Production errors: if a node has a high error rate or a dominant error, call it out — "NotFoundError dominates — likely missing validation" or "clean production profile, no action needed"
-- For healthy nodes: what makes them a good example
-
-Tone: opinionated, direct, like a senior engineer's code review.
-
-Good examples:
-- "God service — owns auth, queries, and scheduling"
-- "Split candidate: query building vs execution"
-- "Regression risk — complex and actively changing"
-- "Stable, well-scoped — good extraction example"
-- "Bottleneck: every controller depends on this"
-- "Thin wrapper, could inline into parent service"
-
-Bad examples (NEVER do these):
-- "CRITICAL: 35 edges, 286 branches" (just restating numbers)
-- "High coupling (45 edges) with active churn" (restating with labels)
-- "Complex with 332 cognitive complexity" (restating)
-
-NODES AND METRICS:
-${nodeDescriptions}
-
-Return a summary for each node. Use ONLY the node name as the key (e.g., "ProjectService", not "ProjectService (service)").`;
-
-    const jsonSchema = JSON.stringify({
-        type: 'object',
-        properties: {
-            summaries: {
-                type: 'object',
-                additionalProperties: { type: 'string' },
-            },
-        },
-        required: ['summaries'],
-        additionalProperties: false,
-    });
-
-    console.log(
-        `Calling Claude to summarize health for ${graph.nodes.length} nodes...`,
-    );
-    const result = execSync(
-        `echo ${escapeShellArg(prompt)} | claude -p --dangerously-skip-permissions --output-format json --json-schema ${escapeShellArg(jsonSchema)} --model sonnet`,
-        { encoding: 'utf-8', maxBuffer: 10 * 1024 * 1024 },
-    );
-
-    const parsed = JSON.parse(result);
-    const output = parsed.structured_output ?? parsed;
-    const summaries: Record<string, string> = output.summaries;
-
-    const summaryCount = Object.keys(summaries).length;
-    console.log(
-        `Got health summaries for ${summaryCount} nodes (${graph.nodes.length} requested).`,
-    );
-
-    const cacheData = {
-        _hash: hash,
-        _generatedAt: new Date().toISOString(),
-        summaries,
-    };
-    fs.writeFileSync(
-        HEALTH_SUMMARY_CACHE,
-        JSON.stringify(cacheData, null, 2) + '\n',
-    );
-    console.log(`Cached to ${HEALTH_SUMMARY_CACHE}`);
-
-    return summaries;
-}
-
-function escapeShellArg(arg: string): string {
-    return "'" + arg.replace(/'/g, "'\\''") + "'";
-}
-
-// ---------------------------------------------------------------------------
-// Main
-// ---------------------------------------------------------------------------
-
-const args = process.argv.slice(2);
-const jsonOnly = args.includes('--json');
-const outIdx = args.indexOf('--out');
-const outDir = outIdx >= 0 ? args[outIdx + 1] : null;
-const forceRefresh = args.includes('--refresh');
-const includeSentry = args.includes('--sentry');
-
-const services = parseServiceRepository();
-const controllers = parseControllers();
-const routers = parseRouters();
-
-const eeParsed = parseEeIndex();
-const eeControllers = parseEeControllers();
-const eeData = {
-    services: eeParsed.services,
-    controllers: eeControllers,
-    modelNames: eeParsed.modelNames,
-    clientNames: eeParsed.clientNames,
-};
-
-const graph = buildGraph(services, controllers, routers, eeData);
-collectGitActivity(graph.nodes, true);
-
-const domains = classifyDomains(graph, forceRefresh);
-const nodeToDomain: Record<string, string> = {};
-for (const [domain, nodeIds] of Object.entries(domains)) {
-    for (const id of nodeIds) {
-        nodeToDomain[id] = domain;
-    }
-}
-for (const node of graph.nodes) {
-    if (nodeToDomain[node.id]) {
-        node.domain = nodeToDomain[node.id];
-    }
-}
-
-const summaries = summarizeGitActivity(graph, forceRefresh);
-for (const node of graph.nodes) {
-    const gs = summaries[node.id] || summaries[`${node.id} (${node.type})`];
-    if (gs) {
-        node.gitSummary = gs;
-    }
-}
-
-const healthSummaries = summarizeHealthScores(graph, forceRefresh);
-for (const node of graph.nodes) {
-    const hs = healthSummaries[node.id] || healthSummaries[`${node.id} (${node.type})`];
-    if (hs) {
-        node.healthSummary = hs;
-    }
-}
-
-runDuplicationAnalysis(graph, forceRefresh);
-summarizeDuplication(graph, forceRefresh);
-
-if (includeSentry) {
-    collectSentryActivity(graph, forceRefresh);
-}
-
-if (jsonOnly) {
-    process.stdout.write(JSON.stringify(graph, null, 2));
-    process.exit(0);
-}
-
-const html = generateHtml(graph);
-
-const outputPath = outDir
-    ? path.resolve(outDir, 'dep-graph.html')
-    : path.join('/tmp', 'lightdash-dep-graph.html');
-
-if (outDir) {
-    fs.mkdirSync(outDir, { recursive: true });
-}
-fs.writeFileSync(outputPath, html);
-
-const eeNodeCount = graph.nodes.filter((n) => n.ee).length;
-const eeLabel = eeNodeCount > 0 ? ` (+${eeNodeCount} EE)` : '';
-console.log(
-    `Generated: ${graph.stats.controllers} controllers, ` +
-        `${graph.stats.routers} routers, ` +
-        `${graph.stats.services} services, ${graph.stats.models} models, ` +
-        `${graph.stats.clients} clients (${graph.stats.totalEdges} edges)${eeLabel}`,
-);
-console.log(`Written to: ${outputPath}`);
-
-// Open in browser
-if (!args.includes('--publish')) {
-    const platform = process.platform;
-    const openCmd =
-        platform === 'darwin'
-            ? 'open'
-            : platform === 'win32'
-              ? 'start'
-              : 'xdg-open';
-    try {
-        execSync(`${openCmd} "${outputPath}"`);
-    } catch {
-        // silent — might be running headless
-    }
-}
-
-// Publish to GitHub Pages
-if (args.includes('--publish')) {
-    const tmp = fs.mkdtempSync(path.join(os.tmpdir(), 'dep-graph-'));
-    execSync('git clone --depth 1 git@github.com:charliedowler/lightdash-dep-graph.git .', { cwd: tmp, stdio: 'inherit' });
-    fs.writeFileSync(path.join(tmp, 'index.html'), html);
-    execSync('git add index.html', { cwd: tmp });
-    const dateStr = new Date().toISOString().slice(0, 10);
-    const commitMsg = `Update dep-graph — ${dateStr}`;
-    try {
-        execSync(`git commit -m "${commitMsg}"`, { cwd: tmp, stdio: 'inherit' });
-        execSync('git push', { cwd: tmp, stdio: 'inherit' });
-        console.log('Published to https://charliedowler.github.io/lightdash-dep-graph/');
-    } catch {
-        console.log('No changes to publish.');
-    }
-    fs.rmSync(tmp, { recursive: true });
-}
+main();
